Fitting Univariate Partial Least Squares Models to Free Parameters in ABC

Function returnPLSModel fits a PLS regression (using plsr) individually to each freely varying parameter of a model, unlike a true multivariate PLS regression. A secondary step than limits the number of components to those that explain some minimum cumulative percentage of variance (see argument variance.cutoff). For ABC, this seems to result in much better results, without one parameter dominating the combined variance.

returnPLSModel(trueFreeValuesMatrix, summaryValuesMatrix, validation = "CV",
  scale = TRUE, variance.cutoff = 95, ...)

PLSTransform(summaryValuesMatrix, pls.model)

Arguments

trueFreeValuesMatrix
Matrix of true free values from simulations.

summaryValuesMatrix
Matrix of summary statistics from simulations.

validation
Character argument controlling what validation procedure is used by plsr. Default is "CV" for cross-validation.

scale
This argument is passed to plsr. It may be a numeric vector, or logical. If numeric vector, the input is scaled by dividing each variable with the corresponding element of scale. If scale = TRUE, the inpus is scaled by dividing each variable by its sample standard deviation. If cross-validation is selected (the default for returnPLSModel), scaling by the standard deviation is done for every segment.

variance.cutoff
Minimum threshold percentage of variance explained for the number of components included in the final PLS model fit. This value is a percentage and must be between 0 and 100. Default is 95 percent.

...
Additional arguments, passed to plsr. In particular, if the number of observations is less than 10 (such as for the example below), it may be necessary to pass a revised segments argument, which plsr passes to mvrCv, as the default segments value is 10 (and there cannot be more segments than observations). Note that this should be an atypical situation outside of examples, as the number of observations should often be much greater than 10.

pls.model
Output from returnPLSModel.

Value

Function returnPLSModel returns a PLS model, and function PLSTransform returns transformed summary statistics.

Details

Function PLSTransform uses results from a Partial Least Squares (PLS) model fit with returnPLSModel to transform summary values.

See also

Function returnPLSModel effectively wraps function plsr from package mvr.

Examples

set.seed(1)
data(simRunExample)

#> Warning: data set ‘simRunExample’ not found

# example simulation

nSimulations<-6

simDataParallel<-parallelSimulateWithPriors(
  nrepSim=nSimulations, multicore=FALSE, coreLimit=1,
  phy=simPhy,
  intrinsicFn=brownianIntrinsic,
  extrinsicFn=nullExtrinsic,
  startingPriorsFns="normal",
  startingPriorsValues=matrix(c(mean(simChar[,1]), sd(simChar[,1]))),
  intrinsicPriorsFns=c("exponential"),
  intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE),
  extrinsicPriorsFns=c("fixed"),
  extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
  timeStep=1e-04,
  checkpointFile=NULL, checkpointFreq=24,
  verbose=FALSE,
  freevector=NULL, taxon.df=NULL,
  niter.brown=25, niter.lambda=25, niter.delta=25, niter.OU=25, niter.white=25)

#> You have only 2 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 5.5e-06 or at the very least 9.2e-05

#> Using 1 core(s) for simulations 
#> 

#> Doing simulations: 

#> .

#> .

#> .

#> .

#> .

#> .

nParFree<-sum(attr(simDataParallel,"freevector"))

#separate the simulation results: 'true' generating parameter values from the summary values
trueFreeValuesMat<-simDataParallel[,1:nParFree]
summaryValuesMat<-simDataParallel[,-1:-nParFree]

PLSmodel<-returnPLSModel(trueFreeValuesMatrix=trueFreeValuesMat,
                summaryValuesMatrix=summaryValuesMat,
                validation="CV", scale=TRUE, variance.cutoff=95 , segments = nSimulations)

#> Warning: in practice, doing PLS works best if you do each free parameter separately, so one parameter does not dominate

PLSmodel

#> Partial least squares regression , fitted with the kernel algorithm.
#> Call:
#> plsr(formula = trueFreeValuesMatrix ~ summaryValuesMatrix, ncomp = ncomp.final,     scale = scale, validation = "none", segments = ..1)

PLSTransform(summaryValuesMatrix=summaryValuesMat, pls.model=PLSmodel)

#>              Comp 1     Comp 2      Comp 3
#> result.1 -7.8307352 -0.8322944  3.20924381
#> result.2  3.8478682  8.2667362 -0.02378598
#> result.3  0.8870224 -1.8778328 -7.11135336
#> result.4 -5.2372870 -1.1874246 -0.21676280
#> result.5 13.9481967 -3.4459104  2.74270421
#> result.6 -5.6150651 -0.9232740  1.39995411