TreEvo--abc for comparative methods

A package for applying Approximate Bayesian Computation to estimating parameters of trait evolution in comparative analyses.

Details

Package:	TreEvo

Type:	Package

Version:	0.3.3

Date:	2012-07-02

License:	GPL

Package:

Examples

# example analysis, using data simulated with TreEvo

set.seed(1)
# let's simulate some data, and then try to infer the parameters using ABC
# get a 30-taxon coalescent tree
tree <- rcoal(20)
# get realistic edge lengths
tree$edge.length <- tree$edge.length*20

genRate <- c(0.01)
ancState <- c(10)

#Simple Brownian motion
simChar <- doSimulation(
    phy = tree,
    intrinsicFn = brownianIntrinsic,
    extrinsicFn = nullExtrinsic,
    startingValues = ancState, #root state
    intrinsicValues = genRate,
    extrinsicValues = c(0),
    generation.time = 100000)

#> Warning: You have only 15 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 0.0011

# NOTE: the example analyses below sample too few particles, 
    # over too few steps, with too few starting simulations
    # - all for the sake of examples that reasonably test the functions

# Please set these values to more realistic levels for your analyses!

results <- doRun_prc(
  phy = tree,
  traits = simChar,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 100000,
  nRuns = 2,
  nStepsPRC = 3,
  numParticles = 20,
     nInitialSimsPerParam = 10,
  jobName = "examplerun_prc",
  stopRule = FALSE,
  multicore = FALSE,
  coreLimit = 1
  )

#> The effective timeStep for this tree will be 0.0035, as a proportion of tree height (root to furthest tip)...

#> Rescaling edge lengths relative to maximum tip-to-root distance...

#> The smallest edge length on the input tree is 0.0019

#> (This is 53000 in the same TreeYears units as used for the input generation.time ( = 1e+05))

#> Warning: Tree has rescaled branches shorter than generation.time/TreeYears; no trait evolution will be simulated on these branches, and thus ML summary stat functions may fail or produce unexpected results.

#> Given generation time, a total of 1389 generations are expected to occur over this tree

#> Number of initial simulations to be performed (nInitialSims) not given
#> Number of initial simulations will instead be product of the number of free parameters multiplied by 10
#> ( = 20 initial simulations)

#> Beginning PRC Run 1 out of a total of 2...

#> Number of initial simulations set to 20

#> Doing initial simulations...

#> Initial simulations took 12.908 seconds

#> Finding distribution of distances...

#> Beginning partial rejection control algorithm...

#> 
#> STARTING DATA GENERATION STEP 1

#> Successes   Attempts   Distance   starting_1  intrinsic_1  

#> 1           40         240        10         0.095                   

#> 2           40         230        9.9         0.11                   

#> 3           40         250        8.4         0.078                   

#> 4           40         140        12         0.05                   

#> 5           40         260        12         0.097                   

#> 6           40         200        9.3         0.048                   

#> 7           40         180        9.3         0.052                   

#> 8           40         180        12         0.047                   

#> 9           40         18        11         0.013                   

#> 10           40         39        9         0.011                   

#> 11           40         94        8.6         0.031                   

#> 12           40         240        9.5         0.063                   

#> 13           40         170        9.6         0.047                   

#> 14           40         140        11         0.0035                   

#> 15           40         260        9.6         0.1                   

#> 16           40         190        9.1         0.053                   

#> 17           40         130        10         0.028                   

#> 18           40         230        11         0.073                   

#> 19           40         130        9.9         0.0046                   

#> 20           40         37        9.6         0.016                   

#> 21           40         120        11         0.03                   

#> 22           40         52        12         0.0092                   

#> 23           40         93        10         0.0057                   

#> 24           40         160        11         0.049                   

#> 25           40         140        9.9         0.039                   

#> 
#> STARTING DATA GENERATION STEP 2

#> Successes   Attempts   Distance   starting_1  intrinsic_1  

#> 1           40         45        9.9         0.019                   

#> 2           40         34        9.6         0.011                   

#> 3           40         110        9.2         0.0036                   

#> 4           40         35        11         0.015                   

#> 5           40         180        9.2         0.045                   

#> 6           40         89        9.4         0.02                   

#> 7           40         150        9.3         0.043                   

#> 8           40         200        11         0.053                   

#> 9           40         180        11         0.091                   

#> 10           40         170        9.8         0.044                   

#> 11           40         130        12         0.033                   

#> 12           40         15        12         0.0096                   

#> 13           40         46        9.1         0.013                   

#> 14           40         140        9.3         0.038                   

#> 15           40         7.7        11         0.0098                   

#> 16           40         130        10         0.03                   

#> 17           40         160        9.5         0.04                   

#> 18           40         94        9.7         0.033                   

#> 19           40         96        11         0.023                   

#> 20           40         38        9         0.014                   

#> 21           40         32        11         0.0069                   

#> 22           40         190        9.5         0.059                   

#> 23           40         87        9.1         0.018                   

#> 24           40         80        9.9         0.024                   

#> 
#> STARTING DATA GENERATION STEP 3

#> Successes   Attempts   Distance   starting_1  intrinsic_1  

#> 1           40         100        11         0.029                   

#> 2           40         130        9.1         0.032                   

#> 3           40         97        8.9         0.026                   

#> 4           40         110        8.9         0.028                   

#> 5           40         81        9         0.019                   

#> 6           40         11        9.3         0.012                   

#> 7           40         8.9        9.7         0.0094                   

#> 8           40         96        9.7         0.017                   

#> 9           40         17        9.4         0.012                   

#> 10           40         95        9.8         0.024                   

#> 11           40         41        9.4         0.013                   

#> 12           40         84        9.6         0.0073                   

#> 13           40         120        9.6         0.031                   

#> 14           40         15        9.5         0.011                   

#> 15           40         49        9.4         0.0071                   

#> 16           40         23        11         0.012                   

#> 17           40         14        9.5         0.0079                   

#> 18           40         130        9         0.036                   

#> 19           40         56        11         0.009                   

#> 20           40         58        10         0.018                   

#> Collection of simulation particles under PRC completed in 28.394 seconds...

#> Run finished...
#> ...
#> 

#> Beginning PRC Run 2 out of a total of 2...

#> Number of initial simulations set to 20

#> Doing initial simulations...

#> Initial simulations took 13.293 seconds

#> Finding distribution of distances...

#> Beginning partial rejection control algorithm...

#> 
#> STARTING DATA GENERATION STEP 1

#> Successes   Attempts   Distance   starting_1  intrinsic_1  

#> 1           40         270        10         0.11                   

#> 2           40         68        12         0.025                   

#> 3           40         50        9         0.016                   

#> 4           40         110        9.7         0.027                   

#> 5           40         130        9         0.03                   

#> 6           40         50        10         0.021                   

#> 7           40         83        9.5         0.007                   

#> 8           40         150        12         0.04                   

#> 9           40         180        11         0.038                   

#> 10           40         36        10         0.023                   

#> 11           40         43        8.5         0.011                   

#> 12           40         77        10         0.019                   

#> 13           40         150        12         0.053                   

#> 14           40         140        12         0.035                   

#> 15           40         130        11         0.0039                   

#> 16           40         170        11         0.044                   

#> 17           40         37        10         0.016                   

#> 18           40         270        12         0.077                   

#> 19           40         170        10         0.05                   

#> 20           40         260        10         0.07                   

#> 21           40         41        9.2         0.0073                   

#> 22           40         160        9.7         0.047                   

#> 23           40         220        8.9         0.054                   

#> 24           40         260        11         0.11                   

#> 25           40         170        10         0.048                   

#> 26           40         82        11         0.028                   

#> 27           40         210        10         0.051                   

#> 28           40         86        10         0.023                   

#> 29           40         150        11         0.037                   

#> 
#> STARTING DATA GENERATION STEP 2

#> Successes   Attempts   Distance   starting_1  intrinsic_1  

#> 1           40         26        9.9         0.015                   

#> 2           40         170        9.6         0.049                   

#> 3           40         140        9.5         0.036                   

#> 4           40         140        9.4         0.0052                   

#> 5           40         130        9.9         0.037                   

#> 6           40         97        9         0.024                   

#> 7           40         68        9.5         0.021                   

#> 8           40         190        8.8         0.074                   

#> 9           40         170        11         0.048                   

#> 10           40         47        11         0.02                   

#> 11           40         88        9.9         0.027                   

#> 12           40         27        10         0.012                   

#> 13           40         190        8.9         0.047                   

#> 14           40         26        9.1         0.017                   

#> 15           40         21        11         0.02                   

#> 16           40         110        9.8         0.029                   

#> 17           40         38        12         0.014                   

#> 18           40         160        12         0.046                   

#> 19           40         160        11         0.041                   

#> 20           40         120        10         0.024                   

#> 21           40         66        10         0.024                   

#> 22           40         96        8.9         0.022                   

#> 23           40         97        9.9         0.023                   

#> 24           40         120        11         0.038                   

#> 25           40         160        9.1         0.044                   

#> 26           40         190        11         0.048                   

#> 27           40         160        11         0.044                   

#> 28           40         130        10         0.028                   

#> 29           40         90        9.1         0.022                   

#> 30           40         180        9.9         0.039                   

#> 
#> STARTING DATA GENERATION STEP 3

#> Successes   Attempts   Distance   starting_1  intrinsic_1  

#> 1           45         99        9         0.034                   

#> 2           45         35        10         0.0089                   

#> 3           45         99        9.2         0.035                   

#> 4           45         98        11         0.0057                   

#> 5           45         38        9.9         0.02                   

#> 6           45         50        9.9         0.017                   

#> 7           45         87        8.8         0.024                   

#> 8           45         130        10         0.038                   

#> 9           45         100        10         0.034                   

#> 10           45         110        12         0.029                   

#> 11           45         81        10         0.031                   

#> 12           45         20        11         0.01                   

#> 13           45         72        8.8         0.0056                   

#> 14           45         26        9.3         0.017                   

#> 15           45         120        11         0.028                   

#> 16           45         130        9.9         0.038                   

#> 17           45         120        9.6         0.033                   

#> 18           45         59        9.8         0.0067                   

#> 19           45         53        10         0.015                   

#> 20           45         10        9.9         0.011                   

#> 21           45         110        9.8         0.021                   

#> Collection of simulation particles under PRC completed in 29.86 seconds...

#> Run finished...
#> ...
#> 

#> Function completed in 190 seconds.