Box-Cox Transformation for Simulation Summary Values

boxcoxTransformation Box-Cox transforms summary values from a single simulation, while boxcoxTransformationMatrix Box-Cox transforms summary values from multiple simulations. The output of boxcoxTransformationMatrix needs to be provided as input for boxcoxTransformation.

boxcoxTransformation(summaryValuesVector, boxcoxAddition, boxcoxLambda)

boxcoxTransformationMatrix(summaryValuesMatrix)

Arguments

summaryValuesVector
Vector of summary statistics from a single simulation

boxcoxAddition
Vector of boxcox additions from boxcoxTransformationMatrix

boxcoxLambda
Vector of boxcox lambda values from boxcoxTransformationMatrix

summaryValuesMatrix
Matrix of summary statistics from simulations

Value

boxcoxTransformation returns a vector of Box-Cox transformed summary statistics from a single observation. boxcoxTransformationMatrix returns a matrix of Box-Cox transformed summary statistics with the same dimensions as summaryValuesMatrix.

Author

Brian O'Meara and Barb Banbury

Examples

# \donttest{
set.seed(1)
data(simRunExample)

# example simulation

simDataParallel <- parallelSimulateWithPriors(
  nrepSim = 5, 
  multicore = FALSE, 
  coreLimit = 1, 
  phy = simPhyExample, 
  intrinsicFn = brownianIntrinsic, 
  extrinsicFn = nullExtrinsic, 
  startingPriorsFns = "normal", 
  startingPriorsValues = list(c(
     mean(simCharExample[, 1]), sd(simCharExample[, 1]))), 
  intrinsicPriorsFns = c("exponential"), 
  intrinsicPriorsValues = list(10), 
  extrinsicPriorsFns = c("fixed"), 
  extrinsicPriorsValues = list(0), 
  generation.time = 100000, 
  checkpointFile = NULL, 
  checkpointFreq = 24, 
  verbose = FALSE, 
  freevector = NULL, taxonDF = NULL)

#> You have only 22 timeSteps on the shortest branch in this dataset
#>   but should probably have a lot more if you expect change on this branch.
#> Please consider decreasing timeStep to no more than 0.0016

nParFree <- sum(attr(simDataParallel, "freevector"))

# separate the simulation results: 
  # 'true' generating parameter values from the summary values
summaryValuesMat <- simDataParallel[, -1:-nParFree]

boxTranMat <- boxcoxTransformationMatrix(
  summaryValuesMatrix = summaryValuesMat
  )
boxTranMat

#> $boxcoxAddition
#>  [1]  820.3567514  173.2757586 1058.5052246  822.1376445  173.3177377
#>  [6]    0.0000000 1054.0670108  817.0842612   86.4910552   15.2940079
#> [11] 1056.3217822  816.0956687  214.1616240    0.3225309 1044.3126219
#> [16]  806.2257106  698.8452487   76.5643983  299.6079455  126.5457849
#> [21] 2513.3034468   76.0895956  126.9176477  150.3826149  240.0795927
#> [26]  587.4282023  278.3068619  169.6408380  218.4550809  153.1170977
#> [31]  117.2190723   82.9188048    5.6530314   40.0833362   15.1508767
#> [36]   67.3700311   34.0335196    3.5583108   24.5686553   44.1398616
#> [41]   24.7986628    7.5101863   76.8317995   83.0127865  101.0390482
#> [46]   52.6028684   14.0180076   96.8229831  141.9021552  158.4403693
#> [51]  286.8401361  131.4034344   92.4112963   52.3340218  279.7332816
#> [56]  130.4185258   83.6438120  115.6177488   54.9420112  133.9111507
#> [61]  115.6375060  143.1654979   75.1565064   57.9603437   36.5585596
#> [66]   78.6457458  129.5121814   74.0648895   94.7469529  248.7138617
#> [71]   84.7296322  146.4685054   18.6056868  242.0663317   95.1650606
#> [76]  108.6389147  137.8635953   38.8566572  114.1178671  126.8350002
#> [81]  268.0231667  239.5687915  238.5356445  166.5259375  124.2996721
#> [86]  173.9458305  211.2021522  222.7254744  344.0735043  180.2100347
#> [91]   69.3458312   99.1123440  321.9640215  167.1934481   65.9415223
#> [96]   97.0765537   73.5254988  154.0555378  105.7927183
#> 
#> $boxcoxLambda
#>  [1]   8.4966 -20.0000  -5.2986   9.0427 -20.0000  20.0000  -5.6862   8.0432
#>  [9] -20.0000   2.7490  -4.9472   8.5867 -20.0000  -1.1314  -5.3238  16.9577
#> [17]  -8.8730 -20.0000 -18.8671  12.3515 -20.0000 -15.7266 -16.5515  -7.5757
#> [25]  -5.5634  20.0000 -17.7518  19.2176 -19.3321  18.7493  20.0000   4.0495
#> [33]  -6.4966  14.4366   3.6926 -20.0000  18.1385  -7.7390 -20.0000 -18.2116
#> [41]   2.5506 -12.8399 -20.0000 -18.8832 -20.0000 -11.3672  -4.6105  -7.3300
#> [49] -13.4826 -12.7981 -20.0000 -16.1207  20.0000  -6.3432 -20.0000 -17.3385
#> [57]  19.8030  19.6383  -7.2234 -16.2093  20.0000  10.7878   8.2401  14.3699
#> [65]   3.1293   8.4990  20.0000 -11.0649 -10.6338 -20.0000 -15.6198  14.3642
#> [73]  -8.4748 -20.0000 -17.3171  15.2878  16.3352  -7.6976 -16.0705  19.8496
#> [81] -17.2908 -16.5299 -17.4494 -14.3970 -12.6367 -20.0000 -14.4161 -13.7614
#> [89] -20.0000 -16.2764   4.5174  -8.4032 -20.0000 -17.1941  20.0000  20.0000
#> [97]  -7.5339 -16.2952  20.0000
#> 
#> $boxcoxSummaryValuesMatrix
#>                                                                          
#> result.1 5.165735e+24 1.670898e-45 8.338559e-17 2.042379e+26 1.662825e-45
#> result.2 4.731359e+24 1.659462e-45 7.690344e-17 1.886231e+26 1.659214e-45
#> result.3 5.889470e+24 1.677232e-45 9.446629e-17 2.347516e+26 1.669126e-45
#> result.4 2.595072e+24 1.126652e-46 4.651298e-17 9.832503e+25 1.121893e-46
#> result.5 3.000516e+24 7.678348e-46 5.213361e-17 1.148226e+26 8.049019e-46
#>                                                                                
#> result.1 1.0000000 5.629414e-18 2.400174e+23 1.803423e-39 2159.551 9.723032e-16
#> result.2 0.7289452 5.226008e-18 2.214790e+23 1.794611e-39 2460.232 9.040382e-16
#> result.3 1.0000000 6.437920e-18 2.744715e+23 1.810754e-39 1922.359 1.102423e-15
#> result.4 1.0000000 3.004830e-18 1.255196e+23 1.213965e-40 2711.581 5.663695e-16
#> result.5 0.9691235 3.399690e-18 1.443989e+23 9.198680e-40 2928.814 6.311943e-16
#>                                                                      
#> result.1 9.026753e+24 2.417191e-47 3.597513 7.426805e-17 9.746046e+48
#> result.2 8.276643e+24 2.400041e-47 3.337236 6.856015e-17 1.029562e+49
#> result.3 1.031291e+25 2.424602e-47 3.597513 8.430883e-17 1.125822e+49
#> result.4 4.499616e+24 1.632074e-48 3.232712 4.115853e-17 2.801968e+48
#> result.5 5.232207e+24 1.056664e-47 3.093494 4.633872e-17 3.776749e+48
#>                                                                          
#> result.1 2.536779e-26 2.609602e-39 1.014434e-47 1.748127e+26 9.769964e-69
#> result.2 2.688936e-26 2.278334e-39 8.971017e-48 1.917247e+26 9.795990e-69
#> result.3 2.960307e-26 1.759793e-39 9.484195e-48 2.107144e+26 9.826814e-69
#> result.4 7.741382e-27 7.981787e-41 6.018459e-49 7.510587e+25 6.600693e-70
#> result.5 1.006539e-26 1.987502e-39 4.232555e-48 5.270489e+25 4.479260e-69
#>                                                                          
#> result.1 5.145195e-31 6.360233e-36 3.691772e-17 5.616538e-14 2.349707e+55
#> result.2 4.590632e-31 4.745385e-36 2.851478e-17 5.369136e-14 2.552190e+55
#> result.3 3.374582e-31 3.928060e-36 2.808528e-17 5.534469e-14 2.374593e+55
#> result.4 3.069154e-32 4.111401e-37 1.962455e-17 3.470062e-14 3.808101e+54
#> result.5 2.508402e-31 3.212603e-36 1.265370e-17 7.474462e-14 2.811041e+55
#>                                                                          
#> result.1 4.649182e-44 6.035438e+42 5.941826e-46 8.725696e+40 2.639240e+41
#> result.2 4.861615e-44 6.649021e+42 6.229711e-46 9.108093e+40 2.493237e+41
#> result.3 4.028681e-44 6.674740e+42 5.957551e-46 9.177626e+40 2.485142e+41
#> result.4 4.543456e-45 2.303638e+42 4.590507e-47 3.079563e+40 3.873397e+40
#> result.5 1.879459e-44 1.128208e+42 2.481586e-46 1.643154e+40 2.739869e+41
#>                                                                               
#> result.1 59047694 1.218781e-05 1.278809e+23 24424.85 2.819525e-37 6.438344e+27
#> result.2 59290468 1.004214e-05 1.245412e+23 22428.81 2.830803e-37 4.644059e+27
#> result.3 58963388 1.069863e-05 1.498571e+23 24211.46 2.733566e-37 6.352371e+27
#> result.4 73913047 1.552229e-05 5.479368e+22 29188.82 2.067234e-38 2.800907e+27
#> result.5 43977888 6.088611e-06 3.724118e+22 18058.98 2.698988e-37 1.070019e+27
#>                                                                      
#> result.1 6.205760e-05 1.559278e-28 1.166416e-30 3642.386 6.178864e-12
#> result.2 7.621785e-05 1.698866e-28 1.285215e-30 3704.002 4.913139e-12
#> result.3 6.187123e-05 1.504137e-28 1.124574e-30 3608.857 5.288954e-12
#> result.4 2.965687e-05 1.168491e-29 1.711285e-31 3043.797 2.273396e-12
#> result.5 3.801414e-05 1.335250e-28 1.993377e-30 4251.292 1.078089e-12
#>                    21           22           23           24           25
#> result.1 2.245135e-39 9.212482e-38 1.644718e-41 5.210733e-21 5.830426e-07
#> result.2 2.164848e-39 7.960180e-38 1.337151e-41 3.643763e-21 3.809290e-07
#> result.3 1.714955e-39 6.443351e-38 1.131948e-41 3.358865e-21 3.785644e-07
#> result.4 7.090698e-41 3.360325e-39 5.531424e-43 5.245613e-22 1.425381e-07
#> result.5 1.840763e-39 4.735024e-38 9.293037e-42 2.527101e-21 3.561803e-07
#>                    26           27           28           29           30
#> result.1 1.377961e-15 4.939139e-30 3.948156e-29 3.904951e-50 3.056964e-35
#> result.2 1.535184e-15 4.097234e-30 3.377665e-29 3.732658e-50 2.364588e-35
#> result.3 1.441834e-15 3.611544e-30 3.020324e-29 3.331999e-50 1.980135e-35
#> result.4 6.363156e-16 5.561525e-31 5.229261e-30 2.189190e-51 2.154686e-36
#> result.5 2.210569e-15 1.981576e-30 1.504120e-29 3.712446e-50 1.447167e-35
#>                    31           32           33           34           35
#> result.1 1.553810e+40 5.178160e-12 6.363909e-50 8.637197e-38 1.080871e+39
#> result.2 1.137914e+40 4.972787e-12 6.096722e-50 6.556671e-38 7.760086e+38
#> result.3 1.294647e+40 4.183893e-12 5.400815e-50 5.377390e-38 8.488854e+38
#> result.4 8.184668e+39 2.126886e-12 3.454492e-51 5.026186e-39 4.715878e+38
#> result.5 1.137784e+39 1.748878e-12 5.301294e-50 4.380764e-38 7.485780e+37
#>                    36           37           38           39             
#> result.1 1.599019e+41 1.054565e-13 1.505625e-35 9.197340e+41 3.129375e+23
#> result.2 1.239828e+41 9.887686e-14 1.133155e-35 7.040313e+41 3.018171e+23
#> result.3 1.417123e+41 8.324402e-14 9.743338e-36 8.174911e+41 3.546591e+23
#> result.4 6.313592e+40 4.233298e-14 1.052689e-36 4.484758e+41 1.230603e+23
#> result.5 1.418335e+40 2.949554e-14 7.200594e-36 7.720780e+40 1.129676e+23
#>                                                                      
#> result.1 6.244746e+15 1.242606e+26 139656.1 2.873898e+16 6.664392e+42
#> result.2 6.356017e+15 1.436800e+26 155710.7 3.424421e+16 5.008525e+42
#> result.3 7.778439e+15 2.126142e+26 168740.1 3.679712e+16 6.389986e+42
#> result.4 3.506761e+15 1.158398e+26 110997.7 1.223863e+16 2.880001e+42
#> result.5 2.588100e+15 2.333567e+25 100179.0 1.612839e+16 6.210628e+41
#>                                                                          
#> result.1 7.772165e-22 4.558067e-22 6.394847e-49 2.361777e-31 3.160444e+31
#> result.2 6.051756e-22 3.774308e-22 6.172920e-49 1.669116e-31 2.684288e+31
#> result.3 4.729004e-22 3.116346e-22 5.250112e-49 1.223213e-31 2.945370e+31
#> result.4 1.035912e-22 7.317081e-23 4.474858e-50 1.545122e-32 1.089332e+31
#> result.5 3.812740e-22 1.977954e-22 8.691996e-49 1.454353e-31 6.327198e+30
#>                                                                          
#> result.1 9.922307e-13 1.072381e-48 1.707668e-35 4.969993e+31 2.663765e+35
#> result.2 9.262688e-13 1.033402e-48 1.204943e-35 4.015467e+31 2.207025e+35
#> result.3 4.977759e-13 8.780235e-49 8.874417e-36 4.334133e+31 2.464289e+35
#> result.4 6.017054e-13 6.138804e-50 8.985581e-37 1.794612e+31 9.488431e+34
#> result.5 1.177412e-13 1.138121e-48 1.093049e-35 7.567791e+30 3.990254e+34
#>                                                                          
#> result.1 1.494841e-13 3.482702e-34 2.401194e+42 5.480650e-43 2.450553e-40
#> result.2 1.379504e-13 2.532464e-34 1.870521e+42 5.209747e-43 2.256843e-40
#> result.3 1.054366e-13 2.091777e-34 2.168762e+42 5.292160e-43 2.253538e-40
#> result.4 6.597772e-14 2.353592e-35 9.459498e+41 4.148002e-44 2.013455e-41
#> result.5 3.129335e-14 1.811714e-34 2.182416e+41 2.172588e-43 9.357820e-41
#>                                                                          
#> result.1 1.669444e-42 4.874769e-33 1.382059e-27 4.831463e-46 1.830665e-34
#> result.2 1.490798e-42 4.031232e-33 1.087945e-27 5.477817e-46 1.573149e-34
#> result.3 1.488486e-42 4.125065e-33 1.159340e-27 5.399466e-46 1.476078e-34
#> result.4 1.170259e-43 4.902148e-34 1.665431e-28 2.515184e-47 2.004704e-35
#> result.5 6.604396e-43 1.818978e-33 5.192357e-28 3.817565e-46 6.814895e-35
#>                                                                       
#> result.1 3.080090e-33 1.104663e-51 1.008988e-37 384693606 8.668189e-18
#> result.2 2.694972e-33 1.050981e-51 8.209629e-38 355087319 8.298785e-18
#> result.3 2.530931e-33 9.777943e-52 7.387864e-38 361362822 7.479000e-18
#> result.4 3.839140e-34 6.099666e-53 7.726699e-39 498503901 2.599394e-18
#> result.5 1.114853e-33 8.151067e-52 4.239654e-38 230356791 2.887105e-18
#>                                                                          
#> result.1 4.058995e-51 2.898176e-39 4.165843e+37 3.838744e+40 4.062654e-15
#> result.2 3.873648e-51 2.306839e-39 2.822930e+37 2.872851e+40 3.831305e-15
#> result.3 3.547647e-51 2.022206e-39 3.019458e+37 3.281197e+40 3.378204e-15
#> result.4 2.197287e-52 1.877317e-40 4.117454e+37 2.365848e+40 1.480030e-15
#> result.5 2.827530e-51 1.301586e-39 2.567098e+36 2.926545e+39 1.256812e-15
#>                                   
#> result.1 1.097079e-36 1.817454e+41
#> result.2 8.484423e-37 1.370079e+41
#> result.3 7.526967e-37 1.588740e+41
#> result.4 7.940714e-38 1.155720e+41
#> result.5 4.935015e-37 1.463428e+40
#> 

boxcoxTransformation(
  summaryValuesVector = summaryValuesMat[, 1], 
  boxcoxAddition = boxTranMat$boxcoxAddition, 
  boxcoxLambda = boxTranMat$boxcoxLambda
  )

#>     result.1     result.2     result.3     result.4     result.5 
#> 5.165735e+24 1.605049e-44 9.291003e-17 9.832503e+25 8.714451e-42 

# }