Run Approximate Bayesian Computation for Phylogenetic Comparative Methods

The doRun functions are the main interface for TreEvo users to do Approximate Bayesian Computation (ABC) analysis, effectively wrapping the simulateWithPriors functions to perform simulations, which themselves are wrappers for the the doSimulation functions. The two current doRun functions are doRun_prc which applies a partial rejection control(PRC) ABC analysis over multiple generations of simulations, and doRun_rej which performs a full rejection ('rej') ABC analysis.

doRun_prc(phy, traits, intrinsicFn, extrinsicFn, startingPriorsValues,
  startingPriorsFns, intrinsicPriorsValues, intrinsicPriorsFns,
  extrinsicPriorsValues, extrinsicPriorsFns, generation.time = 1000,
  TreeYears = max(branching.times(phy)) * 1e+06, multicore = FALSE,
  coreLimit = NA, validation = "CV", scale = TRUE, variance.cutoff = 95,
  niter.goal = 5, numParticles = 300, standardDevFactor = 0.2,
  StartSims = 300, epsilonProportion = 0.7, epsilonMultiplier = 0.7,
  nStepsPRC = 5, jobName = NA, stopRule = FALSE, stopValue = 0.05,
  saveData = FALSE, verboseParticles = TRUE, plot = FALSE)

doRun_rej(phy, traits, intrinsicFn, extrinsicFn, startingPriorsValues,
  startingPriorsFns, intrinsicPriorsValues, intrinsicPriorsFns,
  extrinsicPriorsValues, extrinsicPriorsFns,
  TreeYears = max(branching.times(phy)) * 1e+06, generation.time = 1000,
  multicore = FALSE, coreLimit = NA, validation = "CV", scale = TRUE,
  variance.cutoff = 95, niter.goal = 5, standardDevFactor = 0.2,
  StartSims = NA, jobName = NA, abcTolerance = 0.1,
  checkpointFile = NULL, checkpointFreq = 24, savesims = FALSE)

Arguments

phy
A phylogenetic tree, in package ape's phylo format.

traits
Data matrix with rownames identical to phy@tip.label.

intrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, regardless of the trait values of other taxa.

extrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, based on their own ('internal') trait vlaue and the trait values of other taxa.

startingPriorsValues
Matrix with number of columns equal to the number of states (characters) at root and number of rows equal to two (representing two parameters to pass to prior distribution).

startingPriorsFns
Vector containing names of prior distributions to use for root states: can be one of "fixed", "uniform", "normal",

intrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the intrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

intrinsicPriorsFns
Vector containing names of prior distributions to use for intrinsic function parameters: can be one of "fixed", "uniform", "normal",

extrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the extrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

extrinsicPriorsFns
Vector containing names of prior distributions to use for extrinsic function parameters: can be one of "fixed", "uniform", "normal",

generation.time
The number of years per generation. This sets the coarseness of the simulation; if it's set to 1000, for example, the population's trait values change every 1000 years. Note that this is in calender years (see description for argument TreeYears), and not in millions of years (as is typical for dated trees in macroevolutionary studies). Thus, if a branch is 1 million-year time-unit long, then by default 1000 evolutionary changes occur along that branch.

TreeYears
The amount of calender time from the root to the furthest tip. Most trees in macroevolutionary studies are dated with branch lengths in units of millions of years, and thus the default for this argument is max(branching.times(phy)) * 1e6. If your tree has the most recent tip at time zero (i.e., the modern day), this would be the same as the root age of the tree. If your branch lengths are not in millions of years, you should alter this argument. Otherwise, leave this argument alone.

multicore
Whether to use multicore, default is FALSE. If TRUE, one of two suggested packages must be installed, either doMC (for UNIX systems) or doParallel (for Windows), which are used to activate multithreading. If neither package is installed, this function will fail if multicore=TRUE.

coreLimit
Maximum number of cores to be used.

validation
Character argument controlling what validation procedure is used by plsr. Default is "CV" for cross-validation.

scale
This argument is passed to plsr. It may be a numeric vector, or logical. If numeric vector, the input is scaled by dividing each variable with the corresponding element of scale. If scale = TRUE, the inpus is scaled by dividing each variable by its sample standard deviation. If cross-validation is selected (the default for returnPLSModel), scaling by the standard deviation is done for every segment.

variance.cutoff
Minimum threshold percentage of variance explained for the number of components included in the final PLS model fit. This value is a percentage and must be between 0 and 100. Default is 95 percent.

niter.goal
Adjust number of starting points for package geiger to return the best parameter estimates this number of times on average.

numParticles
Number of accepted particles per generation.

standardDevFactor
Standard deviation for mutating states.

StartSims
Number of initial simulations.

epsilonProportion
Sets tolerance for initial simulations.

epsilonMultiplier
Sets tolerance on subsequent generations.

nStepsPRC
Number of generations to run.

jobName
Optional job name.

stopRule
If TRUE, an analysis will be terminated prior to set number of generations if the ratio of all parameters from one generation to the next falls below the stopValue.

stopValue
Threshold value for terminating an analysis prior to nStepesPRC.

saveData
Option to save various run information during the analysis, including summary statistics from analyses, output to external .Rdata and .txt files.

verboseParticles
If TRUE (the default), a large amount of information about parameter estimates and acceptance of particles is output to console via message as doRun_prc runs.

plot
If TRUE, plots distance of each simulation.

abcTolerance
Proportion of accepted simulations.

checkpointFile
Optional file name for checkpointing simulations.

checkpointFreq
Saving frequency for checkpointing.

savesims
Option to save individual simulations, output to a .Rdata file.

Value

The output of these two functions are lists, composed of multiple objects, which differ slightly in their content among the two functions. For doRun_prc, the output is:

input.data

Input variables: jobName, number of taxa, nrepSim, generation.time, treeYears, epsilonProportion, epsilonMultiplier, nStepsPRC, numParticles, standardDevFactor

PriorMatrix

Matrix of prior distributions

particleDataFrame

DataFrame with information from each simulation, including generation, attempt, id, parentid, distance, weight, and parameter states

toleranceVector

Final tolerance vector

phy

Input phylogeny

traits

Input traits

simTime

Processor time for initial simulations

time.per.gen

Processor time for subsequent generations

credibleInt

Credible Interval calculation for each free parameter of the final generation

HPD

Highest Posterior Density calculation each free parameter of the final generation

For doRun_rej, the output is:

input.data

Input variables: jobName, number of taxa, nrepSim, generation.time, treeYears, epsilonProportion, epsilonMultiplier, nStepsPRC, numParticles, standardDevFactor

PriorMatrix

Matrix of prior distributions

phy

Input phylogeny

traits

Input traits

trueFreeValuesANDSummaryValues

Parameter estimates and summary stats from all sims

simTime

Processor time for simulations

abcDistancesRaw

Euclidean distances for each simulation and free parameter

particleDataFrame

DataFrame with information from each simulation, including generation, attempt, id, parentid, distance, weight, and parameter states

credibleInt

Credible Interval calculation for each free parameter of the final generation

HPD

Highest Posterior Density calculation each free parameter of the final generation

Details

Both doRun functions take an input phylogeny (phy), observed trait data set (traits), models (intrinsicFn, extrinsicFn), and priors (startingPriorsValues, startingPriorsFns, intrinsicPriorsValues, intrinsicPriorsFns, extrinsicPriorsValues, extrinsicPriorsFns). Pulling from the priors, it simulates an initial set of simulations (StartSims). This set of simulations is boxcox transformed ,and a PLS regression (see PLSmethods) is performed for each free parameter to determine the most informative summary statistics (using variance.cutoff). The euclidean distance is calculated between each each initial simulation's most informative summary statistics and the input observed data.

Following that step, the approach of the two functions diverge drastically.

For function doRun_rej, those simulations whose most informative summary statistics fall below abcTolerance are kept as accepted 'particles' (simulations runs), describing the posterior distribution of parameters. No additional generations of simulations are performed.

Coversely, function doRun_prc performs an ABC-PRC analysis, which follows a much more complicated algorithm with additional generations. In PRC, tolerance is set based on epsilonProportion and epsilonMultiplier, and the analysis begins with generation 1 and continues through a number of generations equal to nStepsPRC. Single simulation runs ('particles') are accepted if the distance of the most informative summary stats to the original data is less than this tolerance. A generation is complete when enough particles (numParticles) have been accepted. These particles make up the distribution for the next generation. The accepted particles from the final generation describe the posterior distributions of parameters.

References

Sisson et al. 2007, Wegmann et al. 2009

Examples

data(simRunExample)

#> Warning: data set ‘simRunExample’ not found

# NOTE: the example analyses below sample too few particles, 
        # over too few steps, with too few starting simulations
        # - all for the sake of examples that reasonably test the functions

# Please set these values to more realistic levels for your analyses!

resultsPRC<-doRun_prc(
  phy = simPhy,
  traits = simChar,
  intrinsicFn=brownianIntrinsic,
  extrinsicFn=nullExtrinsic,
  startingPriorsFns="normal",
  startingPriorsValues=matrix(c(mean(simChar[,1]), sd(simChar[,1]))),
  intrinsicPriorsFns=c("exponential"),
  intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE),
  extrinsicPriorsFns=c("fixed"),
  extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
  generation.time=1000,
  standardDevFactor=0.2,
  plot=FALSE,
  StartSims=10,
  epsilonProportion=0.7,
  epsilonMultiplier=0.7,
  nStepsPRC=3,
  numParticles=20,
  jobName="examplerun_prc",
  stopRule=FALSE,
  multicore=FALSE,
  coreLimit=1
)

#> Error in proc.time[[3]]: object of type 'builtin' is not subsettable

resultsPRC

#> Error in eval(expr, envir, enclos): object 'resultsPRC' not found

#one should make sure priors are uniform with doRun_rej!

resultsRej<-doRun_rej(
        phy=simPhy,
        traits=simChar,
        intrinsicFn=brownianIntrinsic,
        extrinsicFn=nullExtrinsic,
        startingPriorsFns="normal",
        startingPriorsValues=matrix(c(mean(simChar[,1]), sd(simChar[,1]))),
        intrinsicPriorsFns=c("exponential"),
        intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE), #grep for normal in pkg
        extrinsicPriorsFns=c("fixed"),
        extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
        StartSims=10,
        jobName="examplerun_rej",
        abcTolerance=0.05,
        multicore=FALSE,
        coreLimit=1
        )

#> Number of initial simulations set to 10

#> Setting number of starting points for Geiger optimization to 
#> 10 for Brownian motion
#> 21 for lambda
#> 19 for delta
#> 10 for OU
#> 10 for white noise

#> You have only 0 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 5.5e-06 or at the very least 9.2e-05

#> Using 1 core(s) for simulations 
#> 

#> Doing simulations: 

#> .

#> .

#> .

#> .

#> .

#> .

#> .

#> .

#> .

#> .

#> Initial simulations took 18.89 seconds

resultsRej

#> $input.data
#>                   [,1]              
#> jobName           "examplerun_rej"  
#>                   "30"              
#> generation.time   "1000"            
#> TreeYears         "829493.294081685"
#> StartSims         "10"              
#> standardDevFactor "0.2"             
#> abcTolerance      "0.05"            
#> 
#> $PriorMatrix
#>        StartingStates 1    IntrinsicValue 1 ExtrinsicValue 1
#> shape  "normal"            "exponential"    "fixed"         
#> value1 "9.84029391358401"  "10"             "0"             
#> value2 "0.503057312936562" "10"             "0"             
#> 
#> $phy
#> 
#> Phylogenetic tree with 30 tips and 29 internal nodes.
#> 
#> Tip labels:
#> 	t14, t5, t4, t1, t20, t10, ...
#> 
#> Rooted; includes branch lengths.
#> 
#> $traits
#>     statesmatrix
#> t30    10.723838
#> t19    10.058869
#> t18     9.587847
#> t12    10.109692
#> t26     9.643529
#> t8     10.513024
#> t4      9.496755
#> t27     9.118159
#> t1     10.355957
#> t20    10.272241
#> t10    10.412839
#> t14     9.370420
#> t5      9.385007
#> t24     9.389497
#> t22     9.600095
#> t28     9.357313
#> t15     8.984641
#> t21     9.573952
#> t11     9.225471
#> t7      9.502582
#> t29    10.173694
#> t9     10.351829
#> t17    10.532439
#> t2     10.523331
#> t6     10.692997
#> t25     9.718049
#> t23     9.596306
#> t3      9.554974
#> t13     9.682015
#> t16     9.701454
#> 
#> $trueFreeValuesANDSummaryValues
#>                [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> result.1   9.965857 0.23448168 -69.637225 53.5338524 143.274449 -69.637225
#> result.2   9.401690 0.01685524  13.022467  0.2164638 -22.044935  13.022467
#> result.3   9.714926 0.11924373 -48.403624 12.9972899 100.807248 -48.403624
#> result.4   9.672004 0.09293566 -28.361559  3.4164539  60.723118 -28.246974
#> result.5   8.428314 0.17422423 -61.863290 31.8823286 127.726580 -61.863290
#> result.6  10.375658 0.06624484 -30.239517  3.8721129  64.479033 -30.229503
#> result.7  10.103134 0.06076900 -19.147866  1.8484676  42.295732 -19.147866
#> result.8   9.793162 0.03689122 -13.849806  1.2984263  31.699612 -13.728334
#> result.9  10.652394 0.05873800 -30.806457  4.0212645  65.612915 -30.802013
#> result.10  9.300095 0.01800871   5.152003  0.3658125  -6.304006   5.377513
#>                 [,7]      [,8]       [,9]      [,10]      [,11]     [,12]
#> result.1  53.5338633 1.0000000 145.274449 -68.274449 22.1657320 2.9999990
#> result.2   0.2164639 1.0000000 -20.044935  13.270854  0.1440654 1.6702464
#> result.3  12.9972920 1.0000000 102.807248 -47.668423  6.5360808 2.3765374
#> result.4   3.0621911 0.9960849  62.493949 -27.586817  1.6574031 2.4975928
#> result.5  31.8823353 1.0000000 129.726580 -61.544971 20.3689473 1.7557609
#> result.6   3.7548177 0.9989721  66.459007 -29.851341  2.4162240 1.8024526
#> result.7   1.8484680 1.0000000  44.295732 -18.239566  0.8099902 2.8784861
#> result.8   1.0679316 0.9917767  33.456668 -13.654046  0.9196107 1.5419915
#> result.9   3.9345257 0.9992763  67.604026 -30.735155  4.8219005 0.7924825
#> result.10  0.3235262 0.9958176  -4.755026   5.428932  0.2456394 1.6457161
#>                [,13]      [,14]      [,15]        [,16]      [,17]      [,18]
#> result.1  142.548898 -68.584148 76.0603670 2.224454e+00 143.168296 -81.647425
#> result.2  -20.541709  13.133092  0.2380557 4.827190e-01 -20.266185  -6.896189
#> result.3  101.336846 -47.885469 15.9085021 1.163648e+00 101.770937 -63.201287
#> result.4   61.173633 -27.617300  4.3094147 1.408003e+00  61.234600 -42.296122
#> result.5  129.089943 -61.734883 35.4278036 5.389488e-01 129.469765 -86.405312
#> result.6   65.702681 -29.792625  4.5219603 8.937779e-01  65.585249 -55.959589
#> result.7   42.479132 -18.404614  2.4072297 1.585720e+00  42.809228 -33.958349
#> result.8   33.308092 -13.564470  1.4671960 6.731577e-01  33.128941 -40.081464
#> result.9   67.470309 -30.806457  4.0212645 6.355331e-54  67.612915 -59.812689
#> result.10  -4.857864   5.429039  0.4164576 7.076042e-01  -4.858077 -20.314437
#>               [,19]     [,20]     [,21]     [,22]       [,23]     [,24]
#> result.1  167.29485  7.930779 14.657664 -2.283981 14.00180332  8.687333
#> result.2   17.79238  9.431923 10.099648  8.811233  0.09592107  9.386072
#> result.3  130.40257 12.690876 17.049797  9.123498  4.09367720 12.143675
#> result.4   88.59224  7.831888 10.466342  5.440184  1.01589089  7.685879
#> result.5  176.81062 10.504031 19.688288  3.989143 19.22821314  9.736512
#> result.6  115.91918  8.612517 11.162430  5.931563  2.52606750  8.872428
#> result.7   71.91670 10.116978 11.516178  8.480709  0.58269840  9.981194
#> result.8   84.16293 11.074829 12.471310  9.196991  0.87644801 11.352154
#> result.9  123.62538 10.002609 12.077454  5.068787  3.26590555 10.656473
#> result.10  44.62887  8.653346  9.532976  7.862205  0.23464542  8.691927
#>               [,25]       [,26]      [,27]       [,28]      [,29]      [,30]
#> result.1   1.675620 -4.83422402 -2.0385070  5.41212740 -0.5874428 -8.7119887
#> result.2   9.027932  0.75737320 -0.5667370 -0.99306479  0.4515953 -0.4176998
#> result.3  11.024264 -0.07102499  4.4021888 -6.61759442 -4.7405470 -1.3751543
#> result.4   7.559573 -4.04509940  0.1638603  1.53265256 -3.0485311 -1.6114322
#> result.5  19.099090  5.78846164  6.7422310 10.42059880 -5.4232435  3.9939643
#> result.6  11.162430  5.36990869  1.1190547  1.17941136 -1.0959436 -0.6741951
#> result.7  10.907823 -0.51435816  2.4866754 -1.04978944  0.4134947  0.3492229
#> result.8  10.703672 -3.73378981  1.1228028  0.24004634  0.8992102 -1.0392350
#> result.9   8.081280 -5.06192589 -6.6319242  0.96729593 -0.3028066  1.6746962
#> result.10  8.986185  1.08380898  1.0894159  0.04049432 -0.7019556 -0.2383336
#>                 [,31]        [,32]      [,33]       [,34]       [,35]
#> result.1   0.80843062 -13.68765980  0.4537971  3.33132116 -1.61340105
#> result.2  -0.15340408  -0.31879481 -0.1309781 -0.03155076 -0.08219019
#> result.3  -0.10269482   0.09276145  2.1401874  0.27927188  0.92644186
#> result.4  -0.01882356   0.59836906 -1.0137856  0.57118025 -1.47458751
#> result.5  -0.11332214   4.18071021  5.7851096  6.80961690 -2.20975449
#> result.6   0.79610396   0.36876540 -1.5015853 -1.20826004  0.67340835
#> result.7   0.05409098   1.10779710 -1.4293558  0.54820246 -1.17997157
#> result.8   0.10538033   0.34522584  0.6017245 -0.06172848  0.94257329
#> result.9  -1.67337139   0.34647157 -0.7257541  1.25891688  0.18216458
#> result.10  0.31588015  -0.41971381  0.5532256 -0.61958715 -0.36948426
#>                [,36]       [,37]       [,38]       [,39]       [,40]
#> result.1   6.4093801  3.24454162 -1.10286940  1.82439725 -2.87730379
#> result.2  -0.2862108 -0.04642761 -0.11914172 -0.08503882  0.20233560
#> result.3  -2.2942447  2.76344919  1.98716894  1.50105187 -1.42137432
#> result.4   0.3819509 -0.61183062 -0.71958730 -0.69396043 -0.33625465
#> result.5  -1.6646881  0.52820693  0.54526836  3.75594500 -0.81168342
#> result.6   2.5126638 -1.31251276  0.10808329 -0.10481549  0.81875291
#> result.7  -0.2666208 -1.58211905 -0.82821799  0.82997110  0.57611160
#> result.8  -0.4782787  0.31632398 -0.02958372  0.36531989  0.58576608
#> result.9  -0.8673881 -0.95098462  0.32937365 -0.95517296  0.28595174
#> result.10  0.4621283  0.38931269  0.28877570 -0.11657291 -0.03354647
#>                  [,41]        [,42]       [,43]       [,44]        [,45]
#> result.1   2.689346577 -0.381412139  0.94833398  0.92260757  0.003938373
#> result.2   0.008243348  0.009930093  0.04028923 -0.04399611 -0.091694640
#> result.3  -0.729439078  0.141496138 -1.08167702  0.19199954 -0.834545358
#> result.4  -0.772094785 -0.333791422  0.26411174 -0.33744351  0.567515804
#> result.5   1.306584364  1.179332789 -0.14395051  0.57262389  0.297882557
#> result.6   0.420271462  0.405541340 -0.18691757  0.04206092  0.390165056
#> result.7  -0.047818726  0.305265618  0.21287062  0.02467459 -0.209223684
#> result.8   0.238733265  0.239881049 -0.18619558 -0.37065576 -0.060293920
#> result.9  -0.150083611 -0.230850673  0.36686639  0.06280334  0.169596944
#> result.10 -0.197916058 -0.049591010  0.03600156 -0.07160149  0.017067390
#>                 [,46]       [,47]       [,48]       [,49]       [,50]
#> result.1   0.94650527  0.42523934  1.11048789 -1.68189275 -1.27698944
#> result.2  -0.15133496 -0.01888815 -0.16651173 -0.07316676  0.04802751
#> result.3  -0.59758343 -0.37779030  0.12522130  0.55479980 -0.75554516
#> result.4  -0.04263398 -0.39376795 -0.30059879  0.07304227 -0.32065227
#> result.5  -1.94013911  1.36673255  1.07934260 -0.04384134 -0.45686361
#> result.6   0.01179279  0.03315682  0.70174432  0.15619496  0.12126030
#> result.7   0.21224380  0.12205842  0.00956220  0.14844616  0.15434727
#> result.8   0.04720769 -0.10831037 -0.18430361  0.11032327  0.22360680
#> result.9   0.69625775  0.40295533 -0.48407681 -0.09673815  0.20229388
#> result.10  0.13230601  0.10689576  0.02968514 -0.01790006 -0.03586981
#>                 [,51]        [,52]       [,53]        [,54]     [,55]     [,56]
#> result.1   0.60610260  0.508706928  0.34458935  0.050085453  7.023412  7.014435
#> result.2   0.03884671 -0.002217275  0.03562914  0.015033443  9.518157  9.519563
#> result.3  -0.35442404 -0.088005939  0.02899420  0.011906646 13.244523 13.244393
#> result.4   0.04081126 -0.055168525 -0.24100390  0.053146398  7.638594  7.631085
#> result.5  -0.24660190  0.285896457  0.10490594  0.339721326 10.714499 10.725247
#> result.6  -0.40116620 -0.079534749  0.01241180  0.150587654  9.041391  9.051360
#> result.7   0.04505391  0.004939849 -0.03736051  0.037451177 10.040462 10.039508
#> result.8   0.26793350 -0.001551898  0.03014829 -0.034475231 10.678092 10.671160
#> result.9   0.23493926  0.191645443 -0.17167237  0.007767371  9.674466  9.665070
#> result.10 -0.10259154 -0.077360342 -0.07415704  0.003105256  8.789975  8.791987
#>               [,57]     [,58]     [,59]     [,60]     [,61]     [,62]     [,63]
#> result.1   6.168507  8.069130  5.356012  6.726594  7.457146  8.173633  8.157524
#> result.2   9.496797  9.354431  9.617795  9.786739  9.214035  9.409405  9.421540
#> result.3  14.009132 13.260326 14.960097 15.872827 12.363918 12.321074 12.119182
#> result.4   7.250922  8.513340  6.998724  7.186514  7.579115  8.092839  8.112452
#> result.5  12.492513  9.463194 11.463511  9.930868 15.857687  8.119122  7.678319
#> result.6   9.790284  7.880156  9.758756  9.885018 10.311565  7.415735  7.458692
#> result.7  10.423742 10.152198 10.616046 10.664192 10.231122 10.257947 10.346964
#> result.8  10.490547 11.485511 10.431967 10.620221 10.510382 11.835711 11.832907
#> result.9   7.991341 10.769469  7.606253  6.993573  7.824936 10.941662 10.997276
#> result.10  9.092502  8.555611  9.134761  9.213874  9.182073  8.366286  8.318782
#>               [,64]     [,65]     [,66]     [,67]     [,68]     [,69]     [,70]
#> result.1   7.047080  7.974505  7.220209  9.617412  8.550329  3.211123  7.618663
#> result.2   9.788825  9.431834  9.467122  9.362540  9.145846  8.989223  9.846732
#> result.3  11.872806 11.960527 12.061755 10.944283 15.019647 11.733669 11.399183
#> result.4   7.079728  8.097057  8.048716  8.186393  9.862895  7.919546  7.539174
#> result.5   9.354108  7.012745  6.671781  5.223661 11.155437 18.667879  9.983915
#> result.6   9.369332  7.551599  7.379799  8.510306  7.992810 10.649803  9.166635
#> result.7   9.206160 10.382342 10.434629 10.367152 10.027222 10.539320  9.531631
#> result.8   9.861326 11.834091 11.881002 11.687625 11.356392 10.638565  9.442502
#> result.9  11.093015 10.972188 11.036744 10.654686 11.158028  7.878571 11.182669
#> result.10  8.588742  8.316743  8.270744  8.456534  8.756365  9.072045  8.696842
#>               [,71]     [,72]     [,73]     [,74]     [,75]     [,76]     [,77]
#> result.1  13.341875 13.456909  1.766507  9.765947  8.447029  8.792992  9.676989
#> result.2   9.215968  9.215228  9.035324  9.282186  9.356595  9.858616  9.438706
#> result.3  11.616879 11.595075 11.175604 11.436272 13.172050 11.033657 11.719003
#> result.4   7.487986  7.512489  7.839414  7.899386  7.425926  7.294013  8.431791
#> result.5  15.684119 15.590345 18.685241  4.768819 10.783457 10.606355  9.710721
#> result.6  10.393222 10.363029 10.968644  8.933057  6.476203  9.307086  7.005368
#> result.7   9.716779  9.685686 10.771112  9.990746  9.530820  9.560266 10.838274
#> result.8  10.384579 10.364190 10.854610 11.617201 12.262235  9.477470 11.796601
#> result.9   7.618173  7.633032  7.983091 10.678001 10.608613 11.135067 11.733759
#> result.10  9.384160  9.390934  9.048785  8.640314  8.602518  8.632571  7.903418
#>               [,78]     [,79]     [,80]     [,81]     [,82]     [,83]
#> result.1   9.561639  6.962139  6.935615 10.208608  6.254500 10.342466
#> result.2   9.078964  9.536948  9.538026  9.367833  9.782433  9.362079
#> result.3  15.996354 13.849381 13.836289 11.758332 12.156269 11.745077
#> result.4   9.640608  7.582988  7.575791  8.314259  6.281095  8.295227
#> result.5  13.311983  6.040884  6.030674 10.247513  8.087991 10.340693
#> result.6   7.949563  6.196991  6.197892  7.290294  9.742468  7.297120
#> result.7  10.450897  9.708835  9.711029 10.860122  8.517316 10.862683
#> result.8  11.535020 12.180902 12.185887 11.716159 10.224112 11.709808
#> result.9  10.696383 10.689819 10.693503 11.552507 11.364494 11.574167
#> result.10  8.716262  8.359980  8.359364  7.901168  8.380340  7.886739
#>                [,84]      [,85]      [,86]      [,87]      [,88]      [,89]
#> result.1  51.1832892 51.0165532 39.0114822 38.1697216 36.0359824 33.9690205
#> result.2   0.8487462  0.8459814  0.6468927  0.6329374  0.5975502  0.5632766
#> result.3  24.0077159 23.9295095 18.2980655 17.9033158 16.9023577 15.9328916
#> result.4   3.3718858  3.3609017  2.5699653  2.5145237  2.3739381  2.2377766
#> result.5  39.1785748 39.0509466 29.8609890 29.2167585 27.5833428 26.0012869
#> result.6   3.5897098  3.5780161  2.7359851  2.6769620  2.5272946  2.3823370
#> result.7   2.4802249  2.4721454  1.8903641  1.8495835  1.7461743  1.6460193
#> result.8   2.0787083  2.0719368  1.5843384  1.5501596  1.4634911  1.3795499
#> result.9   3.6581911  3.6462744  2.7881799  2.7280308  2.5755081  2.4277852
#> result.10  1.1033522  1.0997580  0.8409469  0.8228052  0.7768027  0.7322477
#>                [,90]      [,91]      [,92]      [,93]      [,94]      [,95]
#> result.1  32.9317015 24.6912900 22.6314315 31.6155109 21.6151463 22.4088628
#> result.2   0.5460759  0.4094321  0.3752753  0.5242473  0.3584233  0.3715842
#> result.3  15.4463463 11.5812212 10.6150603 14.8289103 10.1383844 10.5106584
#> result.4   2.1694417  1.6265853  1.4908878  2.0827214  1.4239384  1.4762239
#> result.5  25.2073064 18.8995910 17.3228906 24.1997080 16.5450001 17.1525476
#> result.6   2.3095877  1.7316627  1.5871992  2.2172653  1.5159248  1.5715880
#> result.7   1.5957549  1.1964513  1.0966377  1.5319669  1.0473923  1.0858515
#> result.8   1.3374226  1.0027612  0.9191061  1.2839611  0.8778330  0.9100661
#> result.9   2.3536480  1.7646979  1.6174784  2.2595644  1.5448444  1.6015694
#> result.10  0.7098871  0.5322530  0.4878500  0.6815104  0.4659427  0.4830516
#>                [,96]      [,97]      [,98]      [,99]     [,100]     [,101]
#> result.1  20.2401211 20.6856003 18.4260804 17.4315361 14.1562090 13.5270884
#> result.2   0.3356214  0.3430086  0.3055446  0.2890491  0.2347380  0.2243059
#> result.3   9.4934138  9.7023731  8.6426627  8.1760707  6.6398209  6.3447382
#> result.4   1.3333513  1.3626992  1.2138629  1.1483298  0.9325635  0.8911191
#> result.5  15.4925434 15.8335665 14.1041993 13.3427575 10.8357077 10.3541543
#> result.6   1.4194858  1.4507296  1.2922785  1.2225119  0.9928071  0.9486854
#> result.7   0.9807601  1.0023472  0.8928692  0.8446656  0.6859565  0.6554717
#> result.8   0.8219876  0.8400801  0.7483251  0.7079251  0.5749090  0.5493592
#> result.9   1.4465655  1.4784054  1.3169314  1.2458339  1.0117470  0.9667836
#> result.10  0.4363007  0.4459039  0.3972016  0.3757577  0.3051544  0.2915930
#>               [,102]     [,103]     [,104]     [,105]    [,106]    [,107]
#> result.1  13.5685890 13.0918248 13.3503855 11.5090959 9.7915809 9.0523662
#> result.2   0.2249959  0.2170883  0.2213759  0.1908435 0.1623637 0.1501061
#> result.3   6.3642538  6.1405790  6.2618586  5.3982168 4.5926349 4.2459147
#> result.4   0.8938601  0.8624449  0.8794786  0.7581800 0.6450360 0.5963391
#> result.5  10.3860027 10.0209799 10.2189008  8.8094999 7.4948478 6.9290263
#> result.6   0.9516035  0.9181589  0.9362930  0.8071584 0.6867053 0.6348626
#> result.7   0.6574879  0.6343801  0.6469095  0.5576870 0.4744628 0.4386433
#> result.8   0.5510490  0.5316821  0.5421831  0.4674046 0.3976533 0.3676326
#> result.9   0.9697573  0.9356747  0.9541548  0.8225567 0.6998057 0.6469740
#> result.10  0.2924899  0.2822102  0.2877840  0.2480925 0.2110694 0.1951347
#>              [,108]    [,109]     [,110]     [,111]     [,112]
#> result.1  6.9018141 6.8187348 5.15065609 5.87477343 4.56903998
#> result.2  0.1144457 0.1130681 0.08540803 0.09741532 0.07576368
#> result.3  3.2372224 3.1982549 2.41585946 2.75549893 2.14305872
#> result.4  0.4546682 0.4491952 0.33930768 0.38701007 0.30099279
#> result.5  5.2829132 5.2193212 3.94250840 4.49677543 3.49731727
#> result.6  0.4840398 0.4782132 0.36122696 0.41201093 0.32043693
#> result.7  0.3344358 0.3304101 0.24958121 0.28466919 0.22139830
#> result.8  0.2802949 0.2769209 0.20917721 0.23858490 0.18555676
#> result.9  0.4932738 0.4873362 0.36811814 0.41987092 0.32654995
#> result.10 0.1487770 0.1469862 0.11102863 0.12663786 0.09849119
#> attr(,"freevector")
#> [1]  TRUE  TRUE FALSE
#> 
#> $SimTime
#> [1] 18.89
#> 
#> $abcDistances
#>  result.1  result.2  result.3  result.4  result.5  result.6  result.7  result.8 
#> 26.216919  6.372466 14.749356 11.753485 25.946832 11.895094  5.353060  4.496597 
#>  result.9 result.10 
#> 10.102209  6.356291 
#> 
#> $particleDataFrame
#>          generation attempt id parentid distance weight   param1     param2
#> result.8          1       8  1        0 4.496597      1 9.793162 0.03689122
#> 
#> $credibleInt
#>        mean sd LowerCI UpperCI
#> param1   NA NA      NA      NA
#> param2   NA NA      NA      NA
#> 
#> $HPD
#>        mean sd LowerHPD_0.95 UpperHPD_0.95
#> param1   NA NA            NA            NA
#> param2   NA NA            NA            NA
#>