Approximate Bayesian computation for comparative methods-Rejection

Starts the abc-rejection run

doRun_rej(phy, traits, intrinsicFn, extrinsicFn, startingPriorsValues,
  startingPriorsFns, intrinsicPriorsValues, intrinsicPriorsFns,
  extrinsicPriorsValues, extrinsicPriorsFns, startingValuesGuess = c(),
  intrinsicStatesGuess = c(), extrinsicStatesGuess = c(),
  TreeYears = 10000, standardDevFactor = 0.2, StartSims = NA,
  jobName = NA, abcTolerance = 0.1, multicore = FALSE, coreLimit = NA,
  checkpointFile = NULL, checkpointFreq = 24, validation = "CV",
  scale = TRUE, variance.cutoff = 95, savesims = FALSE, niter.goal = 5,
  generation.time = 1)

Arguments

phy
Tree (Phylogenetic tree in phylo format)

traits
data matrix with rownames equal to phy

intrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, regardless of the states of other taxa

extrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, based on the internal state and the states of other taxa

startingPriorsValues
Matrix with ncol=number of states (characters) at root and nrow=2 (two parameters to pass to prior distribution)

startingPriorsFns
Vector containing names of prior distributions to use for root states: can be one of fixed, uniform, normal, lognormal, gamma, exponential

intrinsicPriorsValues
Matrix with ncol=number of parameters to pass to the intrinsic function and nrow=2 (two parameters to pass to prior distribution)

intrinsicPriorsFns
Vector containing names of prior distributions to use for intrinsic function parameters: can be one of fixed, uniform, normal, lognormal, gamma, exponential

extrinsicPriorsValues
Matrix with ncol=number of parameters to pass to the extrinsic function and nrow=2 (two parameters to pass to prior distribution)

extrinsicPriorsFns
Vector containing names of prior distributions to use for extrinsic function parameters: can be one of fixed, uniform, normal, lognormal, gamma, exponential

startingValuesGuess
Optional guess of starting values

intrinsicStatesGuess
Optional guess of intrinsic values

extrinsicStatesGuess
Optional guess of extrinsic values

TreeYears
Unit length of phy

standardDevFactor
Standard deviation for mutating states

StartSims
Number of simulations

jobName
Optional job name

abcTolerance
Proportion of accepted simulations

multicore
If TRUE, initial simulations will be split among coreLimit nodes

coreLimit
Number of cores for initial simulations

checkpointFile
Optional file name for checkpointing simulations

checkpointFreq
Saving frequency for checkpointing

validation
Cross Validation procedure for abc

scale
scale for pls.model.list

variance.cutoff
variance cutoff for pls.model.list

savesims
option to save individual simulations

niter.goal
Adjust number of starting points for Geiger to return the best parameter estimates this number of times on average

generation.time
The number of years per generation. This sets the coarseness of the simulation; if it's set to 1000, for example, the population moves every 1000 years.

Value

input.data

Input variables: jobName, number of taxa, nrepSim, treeYears, epsilonProportion, epsilonMultiplier, nStepsPRC, numParticles, standardDevFactor

PriorMatrix

Matrix of prior distributions

phy

Input phylogeny

traits

Input traits

trueFreeValuesANDSummaryValues

Parameter estimates and summary stats from all sims

simTime

Processor time for simulations

whichVip

Matrix of vip summary statistics for each free parameter

abcDistancesRaw

Euclidean distances for each simulation and free parameter

particleDataFrame

DataFrame with information from each simulation, including generation, attempt, id, parentid, distance, weight, and parameter states

CredInt

Credible Interval calculation for each free parameter of the final generation

HPD

Highest Posterior Density calculation each free parameter of the final generation

Details

This function performs an abc-rejection analysis using an input phylogeny (phy), character data set (traits), models (intrinsicFn, extrinsicFn), and priors (startingPriorsValues, startingPriorsFns, intrinsicPriorsValues, intrinsicPriorsFns, extrinsicPriorsValues, extrinsicPriorsFns). Pulling from the priors, it performs StartSims simulations. This set of simulations are boxcox transformed and a pls regression is performed for each free parameter to determine the most informative summary statistics (vipthresh). The euclidean distance of each initial simulation's vip summary stats to the input data is calculated, and those that fall under the abcTolerance are kept as accepted particles. These describe the posterior distributions of parameters.

References

O'Meara and Banbury, unpublished; Sisson et al. 2007, Wegmann et al. 2009

Examples

data(simData)

#c<-doRun_rej( #make sure priors are uniform with this one!
#	phy=simPhy,
#	traits=simChar,
#	intrinsicFn=brownianIntrinsic,
#	extrinsicFn=nullExtrinsic,
#	startingPriorsFns="normal",
#	startingPriorsValues=matrix(c(mean(char[,1]), sd(char[,1]))),
#	intrinsicPriorsFns=c("exponential"),
#	intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE), #grep for normal in pkg
#	extrinsicPriorsFns=c("fixed"),
#	extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
#	StartSims=1000,
#	jobName="run_c",
#	abcTolerance=0.05,
#	multicore=F,
#	coreLimit=1
#)