Hacked version of Geiger's fitContinuous

This is TreEvo's version of fitContinuous. This version differs from Geigers in using a different optimization routine, lack of bounds, use of pseudoinverse for matrix inversion, and multiple start values.

fitContinuous.hacked(phy, data, data.names = NULL, model = c("BM", "OU",
  "lambda", "kappa", "delta", "EB", "white", "trend"), bounds = NULL,
  meserr = NULL, userstart = NULL, badLnL = 10000)

Arguments

phy
Phylogenetic tree in phylo format

data
data matrix with rownames equal to phy

data.names
Tip names that match phy

model
Choice of models include "BM", "OU", "lambda", "kappa", "delta", "EB", "white", "trend"

bounds
NULL for fitContinuous.hacked

meserr
Measurement error, if any

userstart
Optional user start input

badLnL
Return value for bad likelihood calculations

Value

Returns a matrix of parameter estimates, along with approximate standard errors and 95 log-likelihood of the model (lnl).

Details

See Geiger's fitContinuous for details.

References

O'Meara and Banbury, unpublished; Harmon et al. 2007

Examples

data(simData)
fitContinuous.hacked(simPhy, simChar, model="lambda")

#> $Trait1
#> $Trait1$lnl
#> [1] -5.385279
#> 
#> $Trait1$beta
#> [1] 0.6922748
#> 
#> $Trait1$lambda
#> [1] 1.002285
#> 
#> $Trait1$aic
#> [1] 16.77056
#> 
#> $Trait1$aicc
#> [1] 17.73056
#> 
#> $Trait1$k
#> [1] 3
#> 
#>