Fitting Univariate Partial Least Squares Models to Free Parameters in ABC

Function returnPLSModel fits a PLS regression (using plsr) individually to each freely varying parameter of a model, unlike a true multivariate PLS regression. A secondary step than limits the number of components to those that explain some minimum cumulative percentage of variance (see argument variance.cutoff). For ABC, this seems to result in much better results, without one parameter dominating the combined variance.

returnPLSModel(trueFreeValuesMatrix, summaryValuesMatrix,
  validation = "CV", scale = TRUE, variance.cutoff = 95,
  verbose = TRUE, ...)

PLSTransform(summaryValuesMatrix, pls.model)

Arguments

trueFreeValuesMatrix
Matrix of true free values from simulations.

summaryValuesMatrix
Matrix of summary statistics from simulations.

validation
Character argument controlling what validation procedure is used by plsr. Default is "CV" for cross-validation.

scale
This argument is passed to plsr. It may be a numeric vector, or logical. If numeric vector, the input is scaled by dividing each variable with the corresponding element of scale. If scale = TRUE, the inpus is scaled by dividing each variable by its sample standard deviation. If cross-validation is selected (the default for returnPLSModel), scaling by the standard deviation is done for every segment.

variance.cutoff
Minimum threshold percentage of variance explained for the number of components included in the final PLS model fit. This value is a percentage and must be between 0 and 100. Default is 95 percent.

verbose
If TRUE, helpful warning messages will be made when you make questionable decisions.

...
Additional arguments, passed to plsr. In particular, if the number of observations is less than 10 (such as for the example below), it may be necessary to pass a revised segments argument, which plsr passes to mvrCv, as the default segments value is 10 (and there cannot be more segments than observations). Note that this should be an atypical situation outside of examples, as the number of observations should often be much greater than 10.

pls.model
Output from returnPLSModel.

Value

Function returnPLSModel returns a PLS model, and function PLSTransform returns transformed summary statistics.

Details

Function PLSTransform uses results from a Partial Least Squares (PLS) model fit with returnPLSModel to transform summary values.

See also

Examples

set.seed(1)
simPhy <- rcoal(20)
simPhy$edge.length <- simPhy$edge.length * 20

# example simulation

nSimulations <- 6

simDataParallel <- parallelSimulateWithPriors(
  nrepSim = nSimulations,
  multicore = FALSE,
  coreLimit = 1,
  phy = simPhy,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 10000,
  checkpointFile = NULL,
  checkpointFreq = 24,
  verbose = FALSE,
  freevector = NULL, taxonDF = NULL)

nParFree <- sum(attr(simDataParallel, "freevector"))

#separate the simulation results: 'true' generating parameter values from the summary values
trueFreeValuesMat <- simDataParallel[, 1:nParFree]
summaryValuesMat <- simDataParallel[, -1:-nParFree]

PLSmodel <- returnPLSModel(trueFreeValuesMatrix = trueFreeValuesMat,
          summaryValuesMatrix = summaryValuesMat,
       validation = "CV", scale = TRUE, variance.cutoff = 95 , segments = nSimulations)

#> Warning: in practice, doing PLS works best if you do each free parameter separately, so one parameter does not dominate

PLSmodel

#> Partial least squares regression , fitted with the kernel algorithm.
#> Call:
#> plsr(formula = trueFreeValuesMatrix ~ summaryValuesMatrix, ncomp = ncomp.final,     scale = scale, validation = "none", segments = ..1)

PLSTransform(summaryValuesMatrix = summaryValuesMat, pls.model = PLSmodel)

#>              Comp 1     Comp 2     Comp 3      Comp 4
#> result.1 -4.4948201  9.5830503 -0.1475185  0.38708095
#> result.2 -4.9661623 -4.3990986  2.1503359  2.97063804
#> result.3 10.6226012  0.6834637 -0.3247972  1.43371025
#> result.4  2.3559766 -0.3608285  3.2078569 -2.91035293
#> result.5  0.6378036 -1.6825234 -3.0030789  0.06319329
#> result.6 -4.1553990 -3.8240635 -1.8827982 -1.94426961