Pairwise Effective Sample Size

This function calculates Effective Sample Size (ESS) on results. The ESS algorithm performs best when results are from multiple runs.

pairwiseESS(inputData)

Arguments

inputData
The input dataset can be a single data frame or a list of data frames.

Value

Returns a matrix with ESS values of all pairwise runs.

Author

Brian O'Meara and Barb Banbury

Examples

data(simRunExample)

# this will give a warning
pairwiseESS(resultsBMExample[[1]]$particleDataFrame)

#> Warning: ESS on a single run should be high; consider combining several runs

#>                  ESS
#> starting_1  55.85303
#> intrinsic_1 88.49379

# ESS should be calculated over multiple runs
pairwiseESS(resultsBMExample)

#> Multiple runs of doRun_prc found, extracting particleDataFrame objects

#>             Runs_1_2
#> starting_1    69.000
#> intrinsic_1  128.526

# you can also manually assemble a list of particleDataFrame tables
    # and use this as the input
inputList <- list(
    resultsBMExample[[2]]$particleDataFrame, 
    resultsBMExample[[1]]$particleDataFrame
    )
pairwiseESS(inputList)

#>             Runs_1_2
#> starting_1   69.0000
#> intrinsic_1 103.7477