Simulate data for initial TreEvo analysis

This is a wrapper function for simulateData that allows for multithreading and checkpointing

parallelSimulation(nrepSim, coreLimit, taxon.df, phy, startingPriorsValues,
  intrinsicPriorsValues, extrinsicPriorsValues, startingPriorsFns,
  intrinsicPriorsFns, extrinsicPriorsFns, freevector, timeStep, intrinsicFn,
  extrinsicFn, multicore, checkpointFile = NULL, checkpointFreq = 24,
  niter.brown = 25, niter.lambda = 25, niter.delta = 25, niter.OU = 25,
  niter.white = 25)

Arguments

nrepSim
Number of simulations

coreLimit
Number of cores to be used

taxon.df
object from getTaxonDFWithPossibleExtinction

phy
Tree (Phylogenetic tree in phylo format)

startingPriorsValues
Matrix with ncol=number of states (characters) at root and nrow=2 (two parameters to pass to prior distribution)

intrinsicPriorsValues
Matrix with ncol=number of states (characters) at root and nrow=2 (two parameters to pass to prior distribution)

extrinsicPriorsValues
Matrix with ncol=number of states (characters) at root and nrow=2 (two parameters to pass to prior distribution)

startingPriorsFns
Vector containing names of prior distributions to use for root states: can be one of fixed, uniform, normal, lognormal, gamma, exponential

intrinsicPriorsFns
Vector containing names of prior distributions to use for root states: can be one of fixed, uniform, normal, lognormal, gamma, exponential

extrinsicPriorsFns
Vector containing names of prior distributions to use for root states: can be one of fixed, uniform, normal, lognormal, gamma, exponential

freevector
A vector (length=number of parameters) of free (T) and fixed (F) parameters

timeStep
This value corresponds to the number of discrete time steps

intrinsicFn
Name of intrinsic function characters should be simulated under (as used by doSimulation)

extrinsicFn
Name of extrinsic function characters should be simulated under (as used by doSimulation)

multicore
Whether to use multicore, default is FALSE

checkpointFile
Optional file name for checkpointing simulations

checkpointFreq
Saving frequency for checkpointing

niter.brown
Number of random starts for BM model (min of 2)

niter.lambda
Number of random starts for lambda model (min of 2)

niter.delta
Number of random starts for delta model (min of 2)

niter.OU
Number of random starts for OU model (min of 2)

niter.white
Number of random starts for white model (min of 2)

Value

Returns matrix of trueFreeValues and summary statistics for simulations

References

O'Meara and Banbury, unpublished