Simulate data for initial TreEvo analysis

The simulateWithPriors function pulls parameters from prior distributions and conducts a single simulation of continuous trait evolution (using the doSimulation function), returning useful summary statistics for ABC. parallelSimulateWithPriors is a wrapper function for simulateWithPriors that allows for multithreading and checkpointing. This family of functions is mostly used as internal components, generating simulations within ABC analyses using the doRun functions. See Note below.

simulateWithPriors(phy = NULL, intrinsicFn, extrinsicFn,
  startingPriorsFns, startingPriorsValues, intrinsicPriorsFns,
  intrinsicPriorsValues, extrinsicPriorsFns, extrinsicPriorsValues,
  generation.time = 1000, TreeYears = max(branching.times(phy)) *
  1e+06, timeStep = NULL, giveUpAttempts = 10, verbose = FALSE,
  checks = TRUE, taxonDF = NULL, freevector = NULL)

parallelSimulateWithPriors(nrepSim, multicore, coreLimit, phy, intrinsicFn,
  extrinsicFn, startingPriorsFns, startingPriorsValues, intrinsicPriorsFns,
  intrinsicPriorsValues, extrinsicPriorsFns, extrinsicPriorsValues,
  generation.time = 1000, TreeYears = max(branching.times(phy)) *
  1e+06, timeStep = NULL, checkpointFile = NULL, checkpointFreq = 24,
  verbose = TRUE, checkTimeStep = TRUE, verboseNested = FALSE,
  freevector = NULL, taxonDF = NULL, giveUpAttempts = 10)

Arguments

phy
A phylogenetic tree, in package ape's phylo format.

intrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, regardless of the trait values of other taxa.

extrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, based on their own ('internal') trait vlaue and the trait values of other taxa.

startingPriorsFns
Vector containing names of prior distributions to use for root states: can be one of "fixed", "uniform", "normal",

startingPriorsValues
Matrix with number of columns equal to the number of states (characters) at root and number of rows equal to two (representing two parameters to pass to prior distribution).

intrinsicPriorsFns
Vector containing names of prior distributions to use for intrinsic function parameters: can be one of "fixed", "uniform", "normal",

intrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the intrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

extrinsicPriorsFns
Vector containing names of prior distributions to use for extrinsic function parameters: can be one of "fixed", "uniform", "normal",

extrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the extrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

generation.time
The number of years per generation. This sets the coarseness of the simulation; if it's set to 1000, for example, the population's trait values change every 1000 calender years. Note that this is in calender years (see description for argument TreeYears), and not in millions of years (as is typical for dated trees in macroevolutionary studies). Thus, if a branch is 1 million-year time-unit long, and a user applies the default generation.time = 1000, then 1000 evolutionary changes will be simulated along that branch. See documentation for doSimulation for further details.

TreeYears
The amount of calender time from the root to the furthest tip. Most trees in macroevolutionary studies are dated with branch lengths in units of millions of years, and thus the default for this argument is max(branching.times(phy)) * 1e6. If your tree has the most recent tip at time zero (i.e., the modern day), this would be the same as the root age of the tree. If your branch lengths are not in millions of years, you should alter this argument. Otherwise, leave this argument alone. See documentation for doSimulation for further details.

timeStep
This value corresponds to the lenght of intervals between discrete evolutionary events ('generations') simulated along branches, relative to a rescaled tree where the root to furthest tip distance is 1. For example, timeStep = 0.01 of would mean 100 (i.e., 1 / 0.01) evolutionary changes would be expected to occur from the root to the furthest tip. (Note that the real number simulated will be much less, because simulations start over at each branching node.) Ideally, timeStep (or its effective value, via other arguments) should be as short as is computationally possible. Typically NULL by default and determined internally as follows: timeStep = generation.time / TreeYears. Can be provided a value as an alternative to using arguments generation.time and TreeYears, which would then be overridden. See documentation for doSimulation for further details.

giveUpAttempts
Value for when to stop the analysis if NA values are present.

verbose
If TRUE, gives messages about how the simulation is progessing via message.

checks
If TRUE, checks inputs for consistency. This activity is skipped (checks = FALSE) when run in parallel by parallelSimulateWithPriors, and instead is only checked once. This argument also controls whether simulateWithPriors assigns freevector as an attribute to the output produced.

taxonDF
A data.frame containing data on nodes (both tips and internal nodes) output by various internal functions. Can be supplied as input to spead up repeated calculations, but by default is NULL, which instead forces a calculation from input phy.

freevector
A logical vector (with length equal to the number of parameters), indicating free (TRUE) and fixed (FALSE) parameters.

nrepSim
Number of replicated simulations to run.

multicore
Whether to use multicore, default is FALSE. If TRUE, one of two suggested packages must be installed, either doMC (for UNIX systems) or doParallel (for Windows), which are used to activate multithreading. If neither package is installed, this function will fail if multicore = TRUE.

coreLimit
Maximum number of cores to be used.

checkpointFile
Optional file name for checkpointing simulations

checkpointFreq
Saving frequency for checkpointing

checkTimeStep
If TRUE, warnings will be issued if TimeStep is too short.

verboseNested
Should looped runs of simulateWithPriors be verbose?

Value

Function simulateWithPriors returns a vector of trueFreeValues, the true generating parameters used in the simulation (a set of values as long as the number of freely varying parameters), concatenated with a set of summary statistics for the simulation. Function parallelSimulateWithPriors returns a matrix of such vectors bound together, with each row representing a different simulation. By default, both functions also assign a logical vector named freevector, indicating the total number of parameters and which parameters are freely-varying (have TRUE values), as an attribute of the output.

Note

The simulateWithPriors functions are effectively the engine that powers the doRun functions, while the doSimulation function is the pistons within the simulateWithPriors engine. In general, most users will just drive the car - they will just use doRun, but some users may want to use simulateWithPriors or doSimulation to do various simulations.

Examples

set.seed(1)
tree <- rcoal(20)
# get realistic edge lengths
tree$edge.length <- tree$edge.length*20

# example simulation

# NOTE: the example analyses involve too few simulations, with coarse time-units
    # - all for the sake of examples that reasonably test the functions

simData <- simulateWithPriors(phy = tree,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 100000,
     freevector = NULL,
     giveUpAttempts = 10,
     verbose = TRUE)

#> Warning: You have only 15 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 0.0011

simData

#>                                                                               
#>   9.91456027   0.09773958 -48.82601235   2.17600214 101.65202469 -48.82601235 
#>                                                                               
#>   2.17600214   1.00000000 103.65202469 -47.74347696   0.91484556   2.96769763 
#>                                                                               
#> 101.48695392 -47.81075449   2.85856000   0.05275491 101.62150897 -60.50373088 
#>                                                                               
#> 125.00746177   4.38526151  13.78384634  -1.70704456  26.14843434   2.62540669 
#>                                                                               
#>   1.68757711   0.39190541 -12.90625436  -6.62196861   3.53946627  -0.42478999 
#>                                                                               
#>   1.09789535  -5.60239695   3.04817260   3.06765236  -4.83815631   0.16137558 
#>                                                                               
#>   0.44555702  -2.69249617  -2.30463728  -0.93426348  -0.48114861  -0.05656878 
#>                                     21           22           23           24 
#>   0.05154922  -0.06159203   5.42015358   5.37728641   6.69727309   2.09868351 
#>           25           26           27           28           29           30 
#>  10.19025936   0.94883494   2.26769958   1.31691492  11.53996433   5.59173481 
#>           31           32           33           34           35           36 
#>   5.71988560   0.38577591   1.28414797  12.05516290   0.12430483  -1.23121903 
#>           37           38           39                                        
#>   0.08770950   2.48540327   9.92510114  -3.44553700  -1.54953474   0.15398828 
#>                                                                               
#>  -3.29571498   4.77619827  -3.90198323  -0.94206919  -1.44375360   8.53670093 
#>                                                                               
#>   3.39459505   3.55761777  -1.85221683  -0.69030566  10.11690689  -1.03612800 
#>                                                                               
#>  -2.27730410  -0.89891408   1.53656780   9.33657065  14.28584415  12.30410755 
#>                                                                               
#>  13.24055790   7.49308199  15.60432045   5.79965311   5.47746836   4.07758345 
#>                                                                               
#>  14.54322772   7.78887457   7.88215342   2.62376865   3.25860160  13.99341890 
#>                                                                  
#>   1.28473767  -0.18513396   1.07433308   3.43423875  10.51363164 
#> attr(,"freevector")
#>  starting_1 intrinsic_1 extrinsic_1 
#>        TRUE        TRUE       FALSE 

simDataParallel <- parallelSimulateWithPriors(
  nrepSim = 2, multicore = FALSE, coreLimit = 1,
  phy = tree,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 100000,
  checkpointFile = NULL, checkpointFreq = 24,
  verbose = TRUE, freevector = NULL, taxonDF = NULL)

#> You have only 15 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 0.0011

#> Using 1 core(s) for simulations 
#> 

#> Doing simulations: 

simDataParallel

#>                                                                          
#> result.1 9.909174 0.005462155   7.717182 0.007621134 -11.43436   7.717182
#> result.2 9.191764 0.073686613 -47.507267 1.907159064  99.01453 -47.507267
#>                                                                             
#> result.1 0.007621134 1  -9.434365   7.720672 0.008059484 0.9292744 -9.441344
#> result.2 1.907159064 1 101.014533 -45.928901 0.757887464 3.0000000 97.857802
#>                                                                            
#> result.1   7.717182 0.007621136 3.233899e-08 -9.434364  -7.702194  19.40439
#> result.2 -45.800838 2.828161799 8.798987e-02 97.601676 -57.010861 118.02172
#>                                                                            
#> result.1  9.868202 10.65029 9.263759  0.1331386  9.746328 10.65029 1.314681
#> result.2 13.900713 20.63514 6.479144 18.4396329 14.582751 13.08247 5.260561
#>                                                                         
#> result.1 0.6990565 -0.3558646 -0.348486  0.1914585 -0.5706051 -0.1265474
#> result.2 7.2539782 -1.5592988  6.809025 -7.8088458  4.7210607  3.7968804
#>                                                                          
#> result.1  0.2140116 -0.1814700  0.1174106 -0.03808816 0.0083272 0.1499546
#> result.2 -4.5796027 -0.6192808 -2.5307003 -0.45959733 0.2887927 4.0812016
#>                                                                               
#> result.1 -0.1025625 -0.02795586 -0.06249713 -0.01450282 0.01371602  0.03156435
#> result.2  2.3125520 -0.44148468  0.05142228  1.02256897 0.15352381 -0.17027051
#>                 21        22        23        24        25        26        27
#> result.1  9.968006  9.824204  9.700072  9.899578  9.670864  9.967784  9.803295
#> result.2 12.775843 12.200435 11.242585 13.623844 10.654293 12.817927 16.437291
#>                 28       29       30       31        32        33       34
#> result.1  9.774474 9.732045 10.54359 10.54124  9.743991  9.808444  9.56847
#> result.2 17.296486 7.387883 15.07898 15.14891 14.433873 16.801405 13.98041
#>                 35        36        37        38       39                  
#> result.1  9.722232  9.641742  9.884776  9.751352 9.849880 9.443328 9.414270
#> result.2 14.421111 20.431922 18.450577 14.814338 7.240305 4.475876 5.715617
#>                                                                          
#> result.1 9.312836 9.580333 9.350456 9.680710  9.613339  9.611096 9.554309
#> result.2 5.116830 8.573665 5.585705 8.276641 13.432339 14.711977 4.576260
#>                                                                             
#> result.1 10.41356 10.41327  9.611545  9.691595  9.453763  9.653557  9.579834
#> result.2 13.02204 13.12462 12.338687 14.952942 12.165834 13.334726 19.452588
#>                                                                          
#> result.1  9.826387  9.695199 9.815051 10.49268 10.23414 10.08731 10.21882
#> result.2 17.526910 13.926048 6.689329 21.07581 18.68525 17.36834 18.67402
#>                                                                          
#> result.1  9.991272 10.25486  9.993252  9.937853  9.90978 10.67362 10.6692
#> result.2 15.722880 17.35921 19.442243 19.880995 10.19951 17.13592 17.1732
#>                                                                              
#> result.1  9.876437  9.925294  9.683177  9.790907  9.70365  9.943165  9.807505
#> result.2 16.529058 18.649867 15.794984 15.507496 21.41126 19.374244 15.702628
#>                  
#> result.1 9.884710
#> result.2 7.791281
#> attr(,"freevector")
#>  starting_1 intrinsic_1 extrinsic_1 
#>        TRUE        TRUE       FALSE