Summarize Posterior Distribution for a Free Parameter

This function summarizes the distribution of parameter estimates from the posterior of an ABC analysis using the doRun functions in TreEvo, for all freely varying parameters. Only the final generation is considered. This summary includes the mean, standard deviation and Highest Posterior Density (at a 0.8 alpha) for each parameter.

summarizePosterior(
  particleDataFrame,
  alpha = 0.8,
  coda = FALSE,
  verboseMultimodal = TRUE,
  stopIfFlat = TRUE,
  ...
)

Arguments

particleDataFrame
A particleDataFrame object, as found among the output from doRun functions.

alpha
The threshold used for defining the highest density frequency cut-off. If the highest density interval is applied to a Bayesian MCMC posterior sample, then the interval is effectively calculated for this value as a posterior probability density.

coda
Default is FALSE. If TRUE, unimodal highest density regions will instead be calculated using HPDinterval from package coda, which is similar to the function quantile in that it calculates only a single interval.

verboseMultimodal
If TRUE, the function will print a message indicating when the inferred highest density interval is discontinuous, and thus likely reflects that the supplied data is multimodal.

stopIfFlat
If TRUE (the default), the function will terminate and return an error if the distribution appears to be flat, with a high number of tied values in the posterior. If FALSE, the function will instead attempt to move forward, while returning a warning instead. The latter option is mainly implemented so internal calls to highestDensityInterval do not block otherwise long-running analyses.

...
Additional arguments passed to density. A user may want to mess with this to adjust bandwidth, et cetera.

Value

Returns a list, wherein each element of the list is secondary list containing the weighted mean, standard deviation, and a matrix giving the highest density intervals (e.g. the highest posterior density intervals). Because posterior estimates of parameter values may be multimodal, multiple sets of bounds may be reported for complex posterior distributions, which each constitute one row of the output matrix. See highestDensityInterval for details.

See also

Author

David W Bapst

Examples

# example with output from doRun_prc
data(simRunExample)

# alpha = 0.95
summarizePosterior(
     resultsBMExample[[1]]$particleDataFrame, 
     alpha = 0.95)

#> $starting_1
#> $starting_1$mean
#> [1] 9.772155
#> 
#> $starting_1$sd
#> [1] 0.126498
#> 
#> $starting_1$HPD
#>      LowerBound_alpha=0.95 UpperBound_alpha=0.95
#> [1,]              9.478273              10.01189
#> 
#> 
#> $intrinsic_1
#> $intrinsic_1$mean
#> [1] 0.03650089
#> 
#> $intrinsic_1$sd
#> [1] 0.02126776
#> 
#> $intrinsic_1$HPD
#>      LowerBound_alpha=0.95 UpperBound_alpha=0.95
#> [1,]          -0.005969742            0.07973995
#> 
#> 

# you might be tempted to use alphas like 95%, 
    # but with bayesian statistics
# we often don't sample the distribution well enough to know
    # its shape to exceeding detail. 
    # alpha = 0.8 may be more reasonable.
summarizePosterior(
    resultsBMExample[[1]]$particleDataFrame, 
    alpha = 0.8)

#> $starting_1
#> $starting_1$mean
#> [1] 9.772155
#> 
#> $starting_1$sd
#> [1] 0.126498
#> 
#> $starting_1$HPD
#>      LowerBound_alpha=0.8 UpperBound_alpha=0.8
#> [1,]             9.593231             9.941538
#> 
#> 
#> $intrinsic_1
#> $intrinsic_1$mean
#> [1] 0.03650089
#> 
#> $intrinsic_1$sd
#> [1] 0.02126776
#> 
#> $intrinsic_1$HPD
#>      LowerBound_alpha=0.8 UpperBound_alpha=0.8
#> [1,]          0.004411246           0.06350302
#> 
#> 

# or even better, for coverage purposes, maybe 0.5
summarizePosterior(
    resultsBMExample[[1]]$particleDataFrame, 
    alpha = 0.5)

#> $starting_1
#> $starting_1$mean
#> [1] 9.772155
#> 
#> $starting_1$sd
#> [1] 0.126498
#> 
#> $starting_1$HPD
#>      LowerBound_alpha=0.5 UpperBound_alpha=0.5
#> [1,]             9.728779             9.902075
#> 
#> 
#> $intrinsic_1
#> $intrinsic_1$mean
#> [1] 0.03650089
#> 
#> $intrinsic_1$sd
#> [1] 0.02126776
#> 
#> $intrinsic_1$HPD
#>      LowerBound_alpha=0.5 UpperBound_alpha=0.5
#> [1,]           0.02091435           0.05498529
#> 
#>