TreEvo--abc for comparative methods

A package for applying Approximate Bayesian Computation to estimating parameters of trait evolution in comparative analyses.

Details

Package: 	 TreEvo

Type: 	 Package

Version: 	 0.3.3

Date: 	 2012-07-02

License: 	 GPL

Examples

# let's simulate some data, and then try to infer the parameters using ABC

simPhy<-rcoal(30)

genRate<-c(0.01)
ancState<-c(10)

#Simple Brownian motion
simCharOut<-doSimulation(
        phy=simPhy,
        intrinsicFn=brownianIntrinsic,
        extrinsicFn=nullExtrinsic,
        startingValues=ancState, #root state
        intrinsicValues=genRate,
        extrinsicValues=c(0),
        timeStep=0.0001,
        saveHistory=FALSE)

# clean for use with doRun
simChar<-simCharOut[,"statesmatrix",drop=FALSE]
rownames(simChar)<-simPhy$tip.label[simCharOut$taxonid]

# NOTE: the example analyses below sample too few particles, 
        # over too few steps, with too few starting simulations
        # - all for the sake of examples that reasonably test the functions

# Please set these values to more realistic levels for your analyses!

results<-doRun_prc(
  phy = simPhy,
  traits = simChar,
  intrinsicFn=brownianIntrinsic,
  extrinsicFn=nullExtrinsic,
  startingPriorsFns="normal",
  startingPriorsValues=matrix(c(mean(simChar[,1]), sd(simChar[,1]))),
  intrinsicPriorsFns=c("exponential"),
  intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE),
  extrinsicPriorsFns=c("fixed"),
  extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
  generation.time=1000,
  standardDevFactor=0.2,
  plot=FALSE,
  StartSims=10,
  epsilonProportion=0.7,
  epsilonMultiplier=0.7,
  nStepsPRC=3,
  numParticles=20,
  jobName="examplerun_prc",
  stopRule=FALSE,
  multicore=FALSE,
  coreLimit=1
  )

#> Error in proc.time[[3]]: object of type 'builtin' is not subsettable