Box-Cox Transformation for Simulation Summary Values

boxcoxTransformation Box-Cox transforms summary values from a single simulation, while boxcoxTransformationMatrix Box-Cox transforms summary values from multiple simulations. The output of boxcoxTransformationMatrix needs to be provided as input for boxcoxTransformation.

boxcoxTransformation(summaryValuesVector, boxcoxAddition, boxcoxLambda)

boxcoxTransformationMatrix(summaryValuesMatrix)

Arguments

summaryValuesVector
Vector of summary statistics from a single simulation

boxcoxAddition
Vector of boxcox additions from boxcoxTransformationMatrix

boxcoxLambda
Vector of boxcox lambda values from boxcoxTransformationMatrix

summaryValuesMatrix
Matrix of summary statistics from simulations

Value

boxcoxTransformation returns a vector of Box-Cox transformed summary statistics from a single observation. boxcoxTransformationMatrix returns a matrix of Box-Cox transformed summary statistics with the same dimensions as summaryValuesMatrix.

Examples

set.seed(1)
data(simRunExample)

# example simulation

simDataParallel <- parallelSimulateWithPriors(
  nrepSim = 5, multicore = FALSE, coreLimit = 1,
  phy = simPhy,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 100000,
  checkpointFile = NULL, checkpointFreq = 24,
  verbose = FALSE,
  freevector = NULL, taxonDF = NULL)

#> You have only 15 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 0.0011

nParFree <- sum(attr(simDataParallel, "freevector"))

#separate the simulation results: 'true' generating parameter values from the summary values
summaryValuesMat <- simDataParallel[, -1:-nParFree]

boxTranMat <- boxcoxTransformationMatrix(summaryValuesMatrix = summaryValuesMat)
boxTranMat

#> $boxcoxAddition
#>  [1]  627.4568294   94.5669444  643.9657077  630.1544011   94.8925531
#>  [6]    0.0000000  641.3608511  632.6738339   74.6567514    9.1838812
#> [11]  655.0053039  632.2432879   96.1573984    0.3607775  651.1740467
#> [16]  768.6784304  661.3747562   36.3975960  182.2535937  169.4761395
#> [21] 2978.6894690   96.0814806   74.4562241  197.8212065  324.7425072
#> [26]   57.4445622  245.2298401   16.6840106  246.2664321   63.8531148
#> [31]   46.6700631   76.0025713   23.5952735   31.8119471   31.2859684
#> [36]   64.6070997   48.2478841   10.3320721   15.5580392   11.5043457
#> [41]   23.5788689    5.6567386   15.2755650    5.3319479    4.2614225
#> [46]   64.8852429   75.2958178   46.8983575  157.1585752  174.4198485
#> [51]  113.6242166   85.9017867   86.4446182   58.1268651  183.4924241
#> [56]  111.5671670   90.9154516  207.5281493  168.9645937  203.9999156
#> [61]  185.0385160  125.0616686  110.4763881  158.9068648   21.8193471
#> [66]  212.7759328   33.2539800  114.1452064  137.2416598  189.9111913
#> [71]   52.6711101   53.7382387  112.1688734  156.7732855  160.9510261
#> [76]  119.6107895  193.3699467  155.6005520  191.1048999  199.9796861
#> [81]  137.9452912   93.6969211   53.3808691  136.5416491    0.0000000
#> [86]  108.1950455  203.7185359  213.2910635   38.9417931  119.2339646
#> [91]  119.2480084   14.6371532  210.9063845   62.5682376   62.8622182
#> [96]  221.7435634  182.4713036  216.9367537  170.1220643
#> 
#> $boxcoxLambda
#>  [1]  11.2504 -20.0000  -5.5140  11.9057 -20.0000  20.0000  -5.9008  10.9502
#>  [9] -20.0000   6.1707  -5.3739  11.1629 -20.0000   6.4678  -5.4867  13.0105
#> [17]  -5.7118  -9.8470 -20.0000  17.5040 -20.0000 -17.9635 -20.0000 -17.2487
#> [25] -20.0000  -2.4628 -20.0000   9.8907  20.0000 -11.0364   8.7335 -20.0000
#> [33]  20.0000  20.0000 -20.0000  20.0000  20.0000 -16.8268 -15.9010   7.5587
#> [41]  20.0000   9.2812  -5.6312   1.5907   1.9972 -12.0258  19.7539  -5.5160
#> [49] -20.0000 -20.0000  20.0000 -20.0000 -20.0000   1.6016 -20.0000  20.0000
#> [57]   7.4234 -20.0000 -20.0000 -20.0000  20.0000  18.8943  16.4741  20.0000
#> [65]   0.3057  20.0000  -3.8329 -20.0000 -20.0000  20.0000 -20.0000 -20.0000
#> [73]   9.2813 -20.0000  20.0000  10.6616 -20.0000 -20.0000 -20.0000  20.0000
#> [81] -20.0000 -20.0000 -12.7349 -20.0000   0.5329 -19.1177 -20.0000 -20.0000
#> [89]  10.3960 -20.0000 -20.0000  -6.6573 -20.0000  19.4234   3.3171 -20.0000
#> [97] -20.0000 -20.0000  20.0000
#> 
#> $boxcoxSummaryValuesMatrix
#>                                                                          
#> result.1 2.610815e+31 3.034391e-40 2.778171e-16 1.858427e+33 2.832871e-40
#> result.2 8.348011e+30 2.122971e-41 1.113090e-16 5.591098e+32 1.998988e-41
#> result.3 1.602248e+31 2.764170e-40 1.822040e-16 1.111183e+33 2.593362e-40
#> result.4 2.112611e+31 2.986989e-40 2.299341e-16 1.520747e+33 2.809490e-40
#> result.5 1.635156e+31 2.792542e-40 1.852511e-16 1.150681e+33 2.680556e-40
#>                                                                               
#> result.1 1.0000000 2.263858e-17 4.225950e+30 3.443049e-38 5012536 6.316339e-16
#> result.2 1.0000000 8.499586e-18 1.384104e+30 2.426061e-39 2063008 2.592479e-16
#> result.3 0.9777476 1.440990e-17 2.626699e+30 3.253108e-38 4050918 4.196985e-16
#> result.4 0.8196329 1.885999e-17 3.431366e+30 3.411454e-38 4551432 5.249092e-16
#> result.5 0.7999782 1.483413e-17 2.644165e+30 3.118192e-38 1811008 4.219591e-16
#>                                                                         
#> result.1 1.641134e+31 2.168414e-40 0.003342111 3.110791e-16 2.523126e+37
#> result.2 5.276532e+30 1.507157e-41 0.001432144 1.252113e-16 7.172092e+36
#> result.3 1.012161e+31 1.930937e-40 0.002698923 2.048775e-16 1.531158e+37
#> result.4 1.335422e+31 2.126787e-40 0.002979392 2.587310e-16 2.011878e+37
#> result.5 1.023113e+31 1.992588e-40 0.001573598 2.066830e-16 1.515169e+37
#>                                                                          
#> result.1 5.444036e-17 3.603569e-17 1.812248e-46 2.619322e+39 3.298130e-70
#> result.2 1.956049e-17 8.795176e-18 8.322673e-48 3.213111e+38 2.356405e-71
#> result.3 3.503684e-17 4.472034e-17 1.553350e-46 1.495498e+39 3.154043e-70
#> result.4 4.430030e-17 5.273907e-17 1.873875e-46 2.060085e+39 3.273579e-70
#> result.5 3.473217e-17 7.258341e-17 1.580115e-46 1.359473e+39 3.145095e-70
#>                                                                          
#> result.1 3.975512e-37 2.697558e-39 2.521816e-40 5.969540e-51 4.447732e-05
#> result.2 3.288579e-38 7.126708e-41 2.173755e-41 5.886592e-52 3.671163e-05
#> result.3 4.005783e-37 2.374904e-39 2.080554e-40 8.509389e-51 4.173567e-05
#> result.4 5.190530e-37 2.261952e-39 1.839504e-40 6.596374e-51 4.629396e-05
#> result.5 6.961158e-37 1.781433e-39 1.042056e-40 6.071833e-51 5.146120e-05
#>                                                                          
#> result.1 1.363343e-48 1.548470e+12 7.080067e+47 1.158964e-20 3.408108e+14
#> result.2 1.040314e-49 8.466521e+11 1.012102e+47 3.327598e-21 1.766712e+14
#> result.3 1.100023e-48 1.170598e+12 4.757315e+47 8.642857e-21 4.925458e+14
#> result.4 1.765222e-48 1.146001e+12 5.995863e+47 9.906200e-21 3.092808e+14
#> result.5 1.002494e-48 5.100425e+11 7.657353e+47 1.589066e-20 3.694445e+14
#>                                                                          
#> result.1 2.407209e-38 3.606155e+27 9.096279e+29 1.356039e-30 1.394866e+36
#> result.2 2.175124e-39 4.807326e+26 1.744201e+29 9.640141e-32 2.212876e+35
#> result.3 3.180587e-38 2.762722e+27 1.514740e+30 1.432800e-30 9.756452e+35
#> result.4 2.515095e-38 3.906802e+27 1.095385e+30 1.168755e-30 1.435795e+36
#> result.5 2.352683e-38 2.251536e+27 8.740611e+29 8.724111e-31 1.564003e+36
#>                                                                                
#> result.1 4.876151e+33 6.288424e-18 1.185795e-19 104034483 2.662091e+27  9998769
#> result.2 6.789310e+32 1.141188e-18 1.692196e-20 129364212 4.837150e+26  4323147
#> result.3 4.984442e+33 6.288013e-18 1.526341e-19  59371451 2.777713e+27  8675233
#> result.4 3.785133e+33 1.011455e-17 9.853049e-20 115628065 3.455564e+27  7068428
#> result.5 3.182008e+33 1.153176e-17 7.002176e-20  66990682 4.048643e+27 12367462
#>                    21       22       23           24           25           26
#> result.1 1.195677e-08 79.07604 210.6547 2.773795e-23 1.400445e+38 2.105982e-10
#> result.2 4.158501e-09 95.49332 125.0667 6.449959e-24 9.916277e+36 1.233721e-10
#> result.3 1.247882e-08 75.59608 212.6033 3.862468e-23 1.452620e+38 2.768921e-10
#> result.4 1.482752e-08 68.43089 173.9372 3.863562e-23 1.104920e+38 2.482847e-10
#> result.5 2.092709e-08 52.44000 112.3976 5.747018e-23 6.384191e+37 3.565920e-10
#>                    27           28           29           30           31
#> result.1 3.081969e-45 4.324359e-46 7.020041e+41 2.245161e-40 2.008891e-40
#> result.2 2.154404e-46 3.053259e-47 6.034864e+40 6.655344e-42 5.981121e-42
#> result.3 3.746576e-45 4.981402e-46 7.035738e+41 1.910961e-40 1.708004e-40
#> result.4 4.615445e-45 6.192233e-46 5.891811e+41 1.965929e-40 1.760497e-40
#> result.5 4.619668e-45 6.401174e-46 3.783665e+41 1.446951e-40 1.292389e-40
#>                32           33           34           35           36
#> result.1 861.0346 1.676609e-46 4.740717e+41 7.490755e+14 1.567817e-47
#> result.2 673.4756 1.181419e-47 4.271888e+40 2.681146e+14 1.137591e-48
#> result.3 744.3011 1.751365e-46 5.081388e+41 4.706083e+14 1.852095e-47
#> result.4 813.9401 2.382850e-46 4.271107e+41 6.211032e+14 2.064694e-47
#> result.5 753.8756 2.462500e-46 3.287522e+41 4.836202e+14 2.284297e-47
#>                    37           38           39                          
#> result.1 8.122502e-46 2.219187e-47 6.381041e+45 1.592333e+40 1.759669e+34
#> result.2 5.725175e-47 1.563767e-48 6.822598e+44 1.517623e+39 2.132628e+33
#> result.3 8.740841e-46 2.146769e-47 6.797532e+45 1.004915e+40 1.190847e+34
#> result.4 1.186601e-45 3.068175e-47 5.764175e+45 1.229023e+40 1.327114e+34
#> result.5 1.213301e-45 3.198956e-47 4.267568e+45 7.508327e+39 7.825413e+33
#>                                                                      
#> result.1 3.361306e+44 2.867376 8.550491e+46 5.716099e-07 1.213440e-42
#> result.2 3.141708e+43 2.778558 9.130811e+45 9.909976e-07 1.035286e-43
#> result.3 2.592052e+44 2.753720 7.335069e+46 8.695832e-07 1.890205e-42
#> result.4 2.655035e+44 2.793396 7.433776e+46 6.970538e-07 2.207482e-42
#> result.5 1.624399e+44 2.682969 5.398811e+46 1.104221e-06 2.484628e-42
#>                                                                          
#> result.1 3.978353e-44 1.006957e+46 1.197136e-36 8.542358e-37 2.272382e+19
#> result.2 3.243341e-45 1.044242e+45 2.544920e-38 1.847213e-38 6.729551e+18
#> result.3 5.407430e-44 9.093537e+45 1.187808e-36 8.419643e-37 1.346072e+19
#> result.4 6.485168e-44 8.770673e+45 1.045710e-36 7.471731e-37 1.864721e+19
#> result.5 7.358684e-44 6.239944e+45 7.606684e-37 5.415665e-37 1.407230e+19
#>                                                                          
#> result.1 3.320897e-45 4.270679e+44 3.289145e+22 6.048688e-47 3.867616e-45
#> result.2 2.556494e-46 4.384394e+43 8.221633e+21 4.569771e-48 2.849660e-46
#> result.3 3.788715e-45 4.150840e+44 1.895601e+22 7.489738e-47 4.361621e-45
#> result.4 5.117041e-45 3.877439e+44 2.643057e+22 8.204591e-47 5.898230e-45
#> result.5 5.607330e-45 3.060351e+44 1.958545e+22 9.355646e-47 6.205785e-45
#>                                                                          
#> result.1 7.730587e-47 2.772048e+46 3.301812e-44 3.888095e-41 9.009986e-24
#> result.2 5.638634e-48 3.021487e+45 1.297359e-45 1.546030e-42 1.005033e-24
#> result.3 7.706033e-47 2.884030e+46 2.263930e-44 2.695802e-41 5.757769e-24
#> result.4 1.101814e-46 2.508299e+46 3.340424e-44 4.425077e-41 1.033251e-23
#> result.5 1.176489e-46 1.873630e+46 3.127605e-44 5.036342e-41 1.305233e-23
#>                                                                      
#> result.1 4.240296e-44 3.537747 2.196149e-40 2.232816e-47 9.323507e-48
#> result.2 2.493491e-45 3.891684 1.640523e-41 1.449871e-48 6.104354e-49
#> result.3 4.352263e-44 3.877826 2.618603e-40 2.346603e-47 9.628489e-48
#> result.4 5.211142e-44 3.451189 2.664349e-40 2.969762e-47 1.223092e-47
#> result.5 6.135814e-44 3.244565 3.975798e-40 2.780716e-47 1.189891e-47
#>                                                                          
#> result.1 4.037158e+17 5.558044e-43 5.556102e-43 5.211131e-10 1.172453e-47
#> result.2 7.882782e+16 1.942953e-44 1.943569e-44 1.900165e-09 7.790703e-49
#> result.3 5.016549e+17 4.395821e-43 4.395861e-43 2.451755e-09 1.146091e-47
#> result.4 3.418306e+17 4.872984e-43 4.874180e-43 9.467749e-10 1.566420e-47
#> result.5 2.196228e+17 3.607829e-43 3.601403e-43 2.113765e-09 1.549285e-47
#>                                                                       
#> result.1 1.370848e+36 1513418.1 4.404457e-48 1.881923e-46 6.830128e-48
#> result.2 9.931892e+34  913824.9 3.086633e-49 1.279320e-47 4.670403e-49
#> result.3 1.829387e+36 1154098.7 4.992907e-48 1.945325e-46 6.434290e-48
#> result.4 1.217791e+36 1354491.9 5.651700e-48 2.648993e-46 9.191904e-48
#> result.5 8.929327e+35 1172439.8 6.089814e-48 2.643165e-46 9.384213e-48
#>                      
#> result.1 1.310945e+45
#> result.2 1.370988e+44
#> result.3 1.435413e+45
#> result.4 1.181904e+45
#> result.5 8.660502e+44
#> 

boxcoxTransformation(summaryValuesVector = summaryValuesMat[, 1],
 boxcoxAddition = boxTranMat$boxcoxAddition, boxcoxLambda = boxTranMat$boxcoxLambda)

#>     result.1     result.2     result.3     result.4     result.5 
#> 2.610815e+31 1.638737e-29 4.369169e-16 1.486819e+33 1.276518e-36