Box-Cox Transformation for Simulation Summary Values

boxcoxTransformation Box-Cox transforms summary values from a single simulation, while boxcoxTransformationMatrix Box-Cox transforms summary values from multiple simulations. The output of boxcoxTransformationMatrix needs to be provided as input for boxcoxTransformation.

boxcoxTransformation(summaryValuesVector, boxcoxAddition, boxcoxLambda)

boxcoxTransformationMatrix(summaryValuesMatrix)

Arguments

summaryValuesVector
Vector of summary statistics from a single simulation

boxcoxAddition
Vector of boxcox additions from boxcoxTransformationMatrix

boxcoxLambda
Vector of boxcox lambda values from boxcoxTransformationMatrix

summaryValuesMatrix
Matrix of summary statistics from simulations

Value

boxcoxTransformation returns a vector of Box-Cox transformed summary statistics from a single observation. boxcoxTransformationMatrix returns a matrix of Box-Cox transformed summary statistics with the same dimensions as summaryValuesMatrix.

Examples

set.seed(1)
data(simRunExample)

# example simulation

simDataParallel <- parallelSimulateWithPriors(
  nrepSim = 5, multicore = FALSE, coreLimit = 1,
  phy = simPhy,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 100000,
  checkpointFile = NULL, checkpointFreq = 24,
  verbose = FALSE,
  freevector = NULL, taxonDF = NULL)

#> You have only 15 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 0.0011

nParFree <- sum(attr(simDataParallel, "freevector"))

#separate the simulation results: 'true' generating parameter values from the summary values
summaryValuesMat <- simDataParallel[, -1:-nParFree]

boxTranMat <- boxcoxTransformationMatrix(summaryValuesMatrix = summaryValuesMat)
boxTranMat

#> $boxcoxAddition
#>  [1]  597.3014015   62.0299474  596.3945200  597.5620753   62.2754256
#>  [6]    0.0000000  588.9158676  600.5393772   52.3129113    8.6505378
#> [11]  596.4550742  601.4410713   61.3414084    0.5898034  597.0691631
#> [16]  674.8148792  534.6386440   68.0197837   49.5460043  338.8822151
#> [21] 1122.6267475   48.3127650   33.3581270  184.1117804  129.1521956
#> [26]  419.0216547  102.9679073   88.7236417   45.3171923   51.4335843
#> [31]   90.4395718   52.7775527   36.5451856   54.0051762   14.9628081
#> [36]   33.6972704   14.6220813    7.8724386   10.7520173    9.4727960
#> [41]    9.9586890    4.2405803   79.2150796  104.3807405  100.7339042
#> [46]  153.9159341   16.5734367  184.7242301  122.9875228  115.1374416
#> [51]   64.8396366   11.6691029   11.1856663  206.6910989   66.4936590
#> [56]   32.7962119  208.3848615  172.2193598   32.1843425   72.9314292
#> [61]   52.3775925  262.4521833  247.3672020  235.8310828  264.6527817
#> [66]   95.3754708  284.3479584  187.9867448  170.9155738   51.5462414
#> [71]   37.0573864   36.7538399  252.7747946  105.8037135   48.3704983
#> [76]  232.3184020  193.3447444   50.9948741   91.7852940   49.9179400
#> [81]    0.0000000    0.0000000    0.0000000   45.1108094   36.5792020
#> [86]   86.1004300   59.6513320   60.7147052   87.4690783   10.4403434
#> [91]    9.5353539  160.6910681   28.7977907   29.1030993  184.4636216
#> [96]  151.2012518   14.3951582   54.3677145   55.3910761
#> 
#> $boxcoxLambda
#>  [1]   9.8838 -20.0000  -4.8242  10.1070 -20.0000 -20.0000  -4.9101  10.7659
#>  [9] -20.0000   8.9524  -5.3331  10.6992 -20.0000   2.4217  -5.2945  13.4502
#> [17]  -5.7274  20.0000 -15.7242  20.0000 -20.0000  16.7058   5.2772 -20.0000
#> [25]  20.0000  20.0000  -9.0151 -20.0000  18.7366   7.8519 -20.0000 -20.0000
#> [33]  -3.3368  20.0000  -9.0641 -20.0000 -17.5239 -18.1969 -13.2215  15.0682
#> [41] -20.0000  -4.4655  20.0000  20.0000  20.0000  20.0000  12.1365  20.0000
#> [49]  20.0000  20.0000  -1.7247  -2.9209  -3.1820  20.0000  20.0000  11.2648
#> [57]  20.0000  20.0000  14.3104  20.0000   3.2095  20.0000  20.0000  20.0000
#> [65]  20.0000  12.5461  20.0000  20.0000  20.0000  12.5437  -4.6283  -4.4654
#> [73]  20.0000  20.0000   6.1529  20.0000  20.0000  20.0000  20.0000   5.9624
#> [81]  -3.9837 -10.3879  -4.0931  17.7555  -7.9282  20.0000  14.9580  18.5126
#> [89]  -9.4126  -1.4757  -1.3497  20.0000  11.8670   9.9108  20.0000  20.0000
#> [97]   7.1801  20.0000   0.9756
#> 
#> $boxcoxSummaryValuesMatrix
#>                                                                          
#> result.1 2.202294e+27 1.378064e-36 3.203439e-14 9.188138e+27 1.277695e-36
#> result.2 9.136318e+26 9.376785e-38 1.574138e-14 3.730365e+27 8.751538e-38
#> result.3 1.629389e+27 1.252677e-36 2.463663e-14 6.771103e+27 1.174912e-36
#> result.4 2.336777e+27 1.386265e-36 3.382425e-14 9.741226e+27 1.281046e-36
#> result.5 1.398289e+27 1.068549e-36 2.173854e-14 5.775372e+27 1.024655e-36
#>                                                                    
#> result.1 1.106717 1.927067e-14 6.609641e+29 4.203870e-35 3518484812
#> result.2 1.000000 9.277057e-15 2.486247e+29 2.951445e-36  839187943
#> result.3 1.059884 1.474873e-14 4.727263e+29 4.003998e-35 3518484812
#> result.4 1.000000 2.033744e-14 6.888554e+29 4.195627e-35 1133104426
#> result.5 1.076312 1.294063e-14 4.032655e+29 3.710984e-35 3518484812
#>                                                                       
#> result.1 1.215865e-15 4.408686e+29 1.705597e-36 0.3877907 1.556800e-15
#> result.2 5.428512e-16 1.662881e+29 1.139386e-37 0.2784337 6.965852e-16
#> result.3 9.017794e-16 3.144438e+29 1.504842e-36 0.3414944 1.151829e-15
#> result.4 1.263854e-15 4.565654e+29 1.730635e-36 0.2910194 1.608643e-15
#> result.5 7.894831e-16 2.683992e+29 1.189466e-36 0.3615994 1.008998e-15
#>                                                                          
#> result.1 7.354016e+37 1.472919e-16 5.917411e+37 1.063243e-28 6.498765e+50
#> result.2 2.133677e+37 5.240625e-17 3.930791e+36 4.490976e-30 7.928218e+49
#> result.3 4.818320e+37 1.001770e-16 5.435528e+37 5.404251e-29 5.410509e+50
#> result.4 6.406779e+37 1.293450e-16 5.246367e+37 1.089172e-28 6.208651e+50
#> result.5 3.984738e+37 8.524541e-17 2.851440e+37 8.464036e-29 4.706790e+50
#>                                                                       
#> result.1 9.802012e-62 2.628625e+29 466831732 3.954588e-46 1.753856e+42
#> result.2 6.812703e-63 2.221252e+28 418115128 6.310493e-47 2.814736e+41
#> result.3 9.231574e-62 2.539450e+29 234550322 8.856685e-46 2.234717e+42
#> result.4 9.717352e-62 2.264089e+29 363325400 6.197421e-46 1.403991e+42
#> result.5 8.365816e-62 1.024771e+29 187639472 6.839592e-46 2.067428e+42
#>                                                                          
#> result.1 2.692268e+52 6.243132e-19 1.031280e-39 1.042548e+31 2.703123e+13
#> result.2 4.524695e+51 2.433169e-19 8.006697e-41 1.673306e+31 3.989659e+13
#> result.3 2.473803e+52 5.551146e-19 6.195392e-40 1.009810e+31 1.647868e+13
#> result.4 2.685320e+52 9.007492e-19 1.297212e-39 9.526099e+30 2.681541e+13
#> result.5 3.944811e+52 5.754742e-19 9.144241e-40 2.035731e+30 4.070999e+13
#>                                                                          
#> result.1 6.417643e-40 2.616845e-35 5.954197e-06 3.785527e+34 2.501454e-11
#> result.2 4.497666e-41 2.541016e-36 4.729094e-06 6.100734e+33 2.011939e-11
#> result.3 6.417320e-40 3.764279e-35 6.455381e-06 4.171278e+34 1.101690e-11
#> result.4 6.913654e-40 3.795274e-35 5.778218e-06 3.944123e+34 2.527868e-11
#> result.5 5.135808e-40 3.067648e-35 7.340943e-06 3.204630e+34 8.192847e-12
#>                                                                          
#> result.1 3.478032e-31 2.484902e-21 6.803530e-17 1.892850e-14 5.503764e+14
#> result.2 2.490740e-32 4.570650e-22 5.902502e-17 3.399768e-15 1.195852e+14
#> result.3 1.665124e-31 5.040996e-21 3.424382e-17 1.474678e-14 4.075967e+14
#> result.4 3.661897e-31 4.534157e-21 4.053720e-17 2.366133e-14 4.534417e+14
#> result.5 2.702561e-31 3.083859e-21 5.655710e-18 1.131093e-14 2.305973e+14
#>                                             21           22           23
#> result.1 1.614309e-20 0.001682891 8.906958e+38 1.267645e+41 6.457010e+40
#> result.2 1.428623e-21 0.002435511 6.235501e+37 1.116009e+40 5.725696e+39
#> result.3 1.915167e-20 0.002160152 6.519746e+38 1.146963e+41 5.958817e+40
#> result.4 1.270182e-20 0.001570108 8.030841e+38 1.266529e+41 6.826986e+40
#> result.5 1.287589e-20 0.002122114 4.566417e+38 8.327307e+40 4.249093e+40
#>                    24           25           26           27           28
#> result.1 1.751278e+44 1.291468e+17 5.436324e+45 2.818275e+42 8.350022e+41
#> result.2 1.880156e+43 7.623058e+16 6.295083e+44 2.741536e+41 7.808388e+40
#> result.3 1.837878e+44 1.446549e+17 5.638712e+45 3.363335e+42 9.331772e+41
#> result.4 1.740777e+44 1.764102e+17 5.780240e+45 2.243228e+42 6.429272e+41
#> result.5 1.418867e+44 1.150347e+16 4.539955e+45 2.361335e+42 7.515327e+41
#>                    29           30           31           32           33
#> result.1 0.0005958138 0.0001197671 5.738982e-05 4.540394e+46 4.519654e+37
#> result.2 0.0004936494 0.0001893457 9.668435e-05 5.421634e+45 3.071977e+36
#> result.3 0.0005765980 0.0002373041 1.227726e-04 4.658312e+46 4.805719e+37
#> result.4 0.0005898348 0.0001645714 8.196052e-05 5.033235e+46 2.783407e+37
#> result.5 0.0006799309 0.0002593095 1.346203e-04 3.213913e+46 4.673688e+37
#>                    34           35           36           37           38
#> result.1 1.847693e+18 5.343474e+46 1.515679e+45 1.661121e+23 2.412016e+38
#> result.2 1.318654e+18 6.283653e+45 1.647415e+44 1.667154e+22 1.661724e+37
#> result.3 1.616202e+18 5.144430e+46 1.533922e+45 2.067812e+23 2.153802e+38
#> result.4 2.615600e+18 5.945341e+46 1.268854e+45 8.468728e+22 1.535628e+38
#> result.5 3.208193e+17 3.421301e+46 1.434736e+45 2.002149e+23 2.516811e+38
#>                39                                                    
#> result.1 542010.2 4.350128e+48 1.389680e+48 5.488445e+47 5.217294e+48
#> result.2 735778.3 5.066629e+47 1.636658e+47 6.466880e+46 6.352690e+47
#> result.3 606605.5 3.358581e+48 1.167983e+48 4.657889e+47 4.888831e+48
#> result.4 570546.2 4.281091e+48 1.409803e+48 5.710747e+47 5.254861e+48
#> result.5 404312.6 2.578842e+48 8.964248e+47 3.544000e+47 3.831245e+48
#>                                                                          
#> result.1 1.923387e+25 2.078757e+49 7.784390e+45 1.285355e+45 2.146817e+22
#> result.2 4.587276e+24 2.608295e+48 8.587257e+44 1.367402e+44 3.901585e+22
#> result.3 1.694083e+25 1.968377e+49 8.119800e+45 1.291166e+45 2.486198e+22
#> result.4 2.122450e+25 2.184566e+49 6.738166e+45 1.081892e+45 2.389171e+22
#> result.5 7.866286e+24 1.582497e+49 5.944240e+45 1.037198e+45 4.742199e+21
#>                                                                         
#> result.1 1.819188e-08 3.506324e-08 2.137300e+48 1.729237e+41 64220043220
#> result.2 3.966672e-08 7.473368e-08 2.648725e+47 1.495834e+40 37202079396
#> result.3 3.099817e-08 5.867146e-08 2.095953e+48 1.660491e+41 57894846082
#> result.4 2.272973e-08 4.358257e-08 2.337847e+48 1.268147e+41 74260409673
#> result.5 3.469183e-08 6.512823e-08 1.484434e+48 1.510461e+41 29028846839
#>                                                                         
#> result.1 4.262210e+47 1.346016e+46 4.436748e+35 1.401295e+40 34802487259
#> result.2 5.127587e+46 1.503967e+45 2.549899e+34 1.097400e+39 56779959784
#> result.3 4.025515e+47 1.327599e+46 5.082081e+35 1.222038e+40 42696686778
#> result.4 4.705030e+47 1.151074e+46 2.344402e+35 9.751241e+39 38503069012
#> result.5 2.726963e+47 1.221901e+46 4.583194e+35 1.331793e+40 19105430387
#>                                                                          
#> result.1 7.380785e-05 2.920382e-11 7.603469e-05 8.220884e+30 5.906900e-14
#> result.2 1.451466e-05 3.163751e-11 8.699254e-05 6.957572e+29 1.457641e-14
#> result.3 5.698097e-05 1.043068e-11 4.967597e-05 1.398296e+31 4.576219e-14
#> result.4 9.612287e-05 3.554648e-11 7.222971e-05 7.717107e+30 5.408282e-14
#> result.5 4.942271e-05 1.326453e-11 5.727498e-05 1.050805e+31 8.604304e-14
#>                                                                        
#> result.1 4.181201e+39 3.799475e+27 1.776911e+34 1.946586e-19 0.01107025
#> result.2 3.873729e+38 5.315777e+26 1.525397e+33 5.425006e-20 0.01040952
#> result.3 5.711407e+39 5.724031e+27 2.520722e+34 1.586796e-19 0.01517829
#> result.4 4.604215e+39 2.690657e+27 1.134579e+34 1.881344e-19 0.01314961
#> result.5 4.629111e+39 4.054851e+27 2.083021e+34 2.915585e-19 0.01522364
#>                                                                        
#> result.1 0.01726312 3.870076e+44 7.556378e+18 5.129459e+15 5.271944e+45
#> result.2 0.01640773 4.368733e+43 1.068784e+18 9.295241e+15 6.026806e+44
#> result.3 0.02335905 4.255468e+44 9.341651e+18 5.084034e+15 5.198121e+45
#> result.4 0.02037604 4.378971e+44 4.200345e+18 7.056361e+15 5.923602e+45
#> result.5 0.02336876 2.821012e+44 1.030775e+19 1.240536e+15 3.378729e+45
#>                                                        
#> result.1 1.323000e+44  8932976186 1.604187e+36 58.07789
#> result.2 1.390923e+43  2142332279 9.135387e+34 64.42334
#> result.3 1.382986e+44 11532968169 1.487458e+36 60.12508
#> result.4 1.078322e+44  4892993355 8.890051e+35 58.93250
#> result.5 1.320054e+44 11941501784 1.936944e+36 53.60564
#> 

boxcoxTransformation(summaryValuesVector = summaryValuesMat[, 1],
 boxcoxAddition = boxTranMat$boxcoxAddition, boxcoxLambda = boxTranMat$boxcoxLambda)

#>     result.1     result.2     result.3     result.4     result.5 
#> 2.202294e+27 1.342415e+00 5.264964e-14 9.741226e+27 7.137467e-28