Discrete Time Character Simulation

This function evolves continuous characters in a discrete time process.

doSimulation(splits, intrinsicFn, extrinsicFn, startingValues, intrinsicValues,
  extrinsicValues, timeStep, saveHistory = FALSE, saveRealParams = FALSE,
  jobName = "")

Arguments

splits
Output from the function getSimulationSplits; is a data frame of branching times, ancestor and descendant vectors

intrinsicFn
Name of intrinsic function characters should be simulated under

extrinsicFn
Name of extrinsic function characters should be simulated under

startingValues
State at the root

intrinsicValues
Vector of values corresponding to the params of the intrinsic model

extrinsicValues
Vector of values corresponding to the params of the extrinsic model

timeStep
This value corresponds to the number of discrete time steps on the shortest branch

saveHistory
Saves the character history throughout the simulation

saveRealParams
Saves intrinsicValues and extrinsicValues as both a matrix and a vector

jobName
Optional name for the job

Value

A data frame of species character (tip) values in the tree.

Details

When saveHistory is TRUE, processor time will increase quite a bit. SaveRealParams is useful for tracking the "real" true values if simulating data for abc runs. It is not useful for empirical abc runs.

References

O'Meara and Banbury, unpublished

Examples

phy<-rcoal(30)

#Simple Brownian motion
char<-doSimulation(
        splits=getSimulationSplits(phy),
        intrinsicFn=brownianIntrinsic,
        extrinsicFn=nullExtrinsic,
        startingValues=c(10), #root state
        intrinsicValues=c(0.01),
        extrinsicValues=c(0),
        timeStep=0.0001,
        saveHistory=FALSE)

#Character displacement model with minimum bound
char<-doSimulation(
        splits=getSimulationSplits(phy),
        intrinsicFn=boundaryMinIntrinsic,
        extrinsicFn=ExponentiallyDecayingPush,
        startingValues=c(10), #root state
        intrinsicValues=c(0.05, 10, 0.01),
        extrinsicValues=c(0, .1, .25),
        timeStep=0.001,
        saveHistory=FALSE)