Example Analysis Output of a Simulated Dataset

Simulated 30-taxon coalescent tree and simulated character data from a Brownian Motion intrinsic model (brownianIntrinsic), with saved generating parameters. Character data was generated under this model using doSimulation. Also includes results from an example analysis.

Format

Loading the simRunExample example dataset adds seven new objects to the namespace:

simPhy

A simulated 30-tip coalescent phylogeny in typical phylo format.

ancState

The starting ancestral value, used for generating the simulated continuous trait data.

genRate

The true rate of trait change under Brownian Motion, used for generating the simulated continuous trait data.

simChar

The output of doSimulation on simPhy, under the model brownianIntrinsic. composed of just the simulated trait values as a one-column matrix with row names indicating tip labels, as desired by doRun functions.

results

The results of doRun_prc, under the generating model of brownianIntrinsic

resultsBound

The results of doRun_prc, under the incorrect model of boundaryMinIntrinsic

The objects results and resultsBound are lists composed of a number of elements (see the documentation for the doRun_prc function for more detail). These elements are respectively input.data, PriorMatrix, particleDataFrame, toleranceVector, phy, traits, simTime, time.per.gen, and postSummary.

Examples

data(simRunExample)

# ...things to do with this data?

###################

# This data was generated using this process:
# NOT RUN {
library(TreEvo)

set.seed(1)
simPhy <- rcoal(20)
# get realistic edge lengths
simPhy$edge.length <- simPhy$edge.length*20

# plot with time axis (root is about ~15 Ma)
plot(simPhy)
axisPhylo()

genRate <- c(0.001)
ancState <- c(10)

#Simple Brownian motion
simChar <- doSimulation(
    phy = simPhy,
    intrinsicFn = brownianIntrinsic,
    extrinsicFn = nullExtrinsic,
    startingValues = ancState, #root state
    intrinsicValues = genRate,
    extrinsicValues = c(0),
    generation.time = 10000
    )

results <- doRun_prc(
    phy = simPhy,
    traits = simChar,
    intrinsicFn = brownianIntrinsic,
    extrinsicFn = nullExtrinsic,
    startingPriorsFns = "normal",
    startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
    intrinsicPriorsFns = c("exponential"),
    intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
    extrinsicPriorsFns = c("fixed"),
    extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
    generation.time = 10000,
    nRuns = 2,
    nStepsPRC = 3,
    numParticles = 20,
    nInitialSimsPerParam = 10,
    jobName = "examplerun_prc",
    stopRule = FALSE,
    multicore = FALSE,
    verboseParticles = TRUE,
    coreLimit = 1
    )

resultsBound <- doRun_prc(
    phy = simPhy,
    traits = simChar,
    intrinsicFn = boundaryMinIntrinsic,
    extrinsicFn = nullExtrinsic,
    startingPriorsFns = "normal",
    startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
    intrinsicPriorsFns = c("exponential", "normal"),
    intrinsicPriorsValues = matrix(c(10, 10, -10, 1), nrow = 2, byrow = FALSE),
    extrinsicPriorsFns = c("fixed"),
    extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
    generation.time = 10000,
    nRuns = 2,
    nStepsPRC = 3,
    numParticles = 20,
    nInitialSimsPerParam = 10,
    jobName = "examplerun_prc_bound",
    stopRule = FALSE,
    multicore = FALSE,
    verboseParticles = TRUE,
    coreLimit = 1
    )

rm(.Random.seed)
save.image(file = "simRunExample.rdata")

# }