Simulate data for initial TreEvo analysis

The simulateWithPriors function pulls parameters from prior distributions and conducts a single simulation of continuous trait evolution (using the doSimulation function), returning useful summary statistics for ABC. parallelSimulateWithPriors is a wrapper function for simulateWithPriors that allows for multithreading and checkpointing. This family of functions is mostly used as internal components, generating simulations within ABC analyses using the doRun functions. See Note below.

simulateWithPriors(phy = NULL, intrinsicFn, extrinsicFn,
  startingPriorsFns, startingPriorsValues, intrinsicPriorsFns,
  intrinsicPriorsValues, extrinsicPriorsFns, extrinsicPriorsValues,
  generation.time = 1000, TreeYears = max(branching.times(phy)) *
  1e+06, timeStep = NULL, giveUpAttempts = 10, verbose = FALSE,
  checks = TRUE, taxonDF = NULL, freevector = NULL)

parallelSimulateWithPriors(nrepSim, multicore, coreLimit, phy, intrinsicFn,
  extrinsicFn, startingPriorsFns, startingPriorsValues, intrinsicPriorsFns,
  intrinsicPriorsValues, extrinsicPriorsFns, extrinsicPriorsValues,
  generation.time = 1000, TreeYears = max(branching.times(phy)) *
  1e+06, timeStep = NULL, checkpointFile = NULL, checkpointFreq = 24,
  verbose = TRUE, checkTimeStep = TRUE, verboseNested = FALSE,
  freevector = NULL, taxonDF = NULL, giveUpAttempts = 10)

Arguments

phy
A phylogenetic tree, in package ape's phylo format.

intrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, regardless of the trait values of other taxa.

extrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, based on their own ('internal') trait vlaue and the trait values of other taxa.

startingPriorsFns
Vector containing names of prior distributions to use for root states: can be one of "fixed", "uniform", "normal",

startingPriorsValues
Matrix with number of columns equal to the number of states (characters) at root and number of rows equal to two (representing two parameters to pass to prior distribution).

intrinsicPriorsFns
Vector containing names of prior distributions to use for intrinsic function parameters: can be one of "fixed", "uniform", "normal",

intrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the intrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

extrinsicPriorsFns
Vector containing names of prior distributions to use for extrinsic function parameters: can be one of "fixed", "uniform", "normal",

extrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the extrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

generation.time
The number of years per generation. This sets the coarseness of the simulation; if it's set to 1000, for example, the population's trait values change every 1000 calender years. Note that this is in calender years (see description for argument TreeYears), and not in millions of years (as is typical for dated trees in macroevolutionary studies). Thus, if a branch is 1 million-year time-unit long, and a user applies the default generation.time = 1000, then 1000 evolutionary changes will be simulated along that branch. See documentation for doSimulation for further details.

TreeYears
The amount of calender time from the root to the furthest tip. Most trees in macroevolutionary studies are dated with branch lengths in units of millions of years, and thus the default for this argument is max(branching.times(phy)) * 1e6. If your tree has the most recent tip at time zero (i.e., the modern day), this would be the same as the root age of the tree. If your branch lengths are not in millions of years, you should alter this argument. Otherwise, leave this argument alone. See documentation for doSimulation for further details.

timeStep
This value corresponds to the lenght of intervals between discrete evolutionary events ('generations') simulated along branches, relative to a rescaled tree where the root to furthest tip distance is 1. For example, timeStep = 0.01 of would mean 100 (i.e., 1 / 0.01) evolutionary changes would be expected to occur from the root to the furthest tip. (Note that the real number simulated will be much less, because simulations start over at each branching node.) Ideally, timeStep (or its effective value, via other arguments) should be as short as is computationally possible. Typically NULL by default and determined internally as follows: timeStep = generation.time / TreeYears. Can be provided a value as an alternative to using arguments generation.time and TreeYears, which would then be overridden. See documentation for doSimulation for further details.

giveUpAttempts
Value for when to stop the analysis if NA values are present.

verbose
If TRUE, gives messages about how the simulation is progessing via message.

checks
If TRUE, checks inputs for consistency. This activity is skipped (checks = FALSE) when run in parallel by parallelSimulateWithPriors, and instead is only checked once. This argument also controls whether simulateWithPriors assigns freevector as an attribute to the output produced.

taxonDF
A data.frame containing data on nodes (both tips and internal nodes) output by various internal functions. Can be supplied as input to spead up repeated calculations, but by default is NULL, which instead forces a calculation from input phy.

freevector
A logical vector (with length equal to the number of parameters), indicating free (TRUE) and fixed (FALSE) parameters.

nrepSim
Number of replicated simulations to run.

multicore
Whether to use multicore, default is FALSE. If TRUE, one of two suggested packages must be installed, either doMC (for UNIX systems) or doParallel (for Windows), which are used to activate multithreading. If neither package is installed, this function will fail if multicore = TRUE.

coreLimit
Maximum number of cores to be used.

checkpointFile
Optional file name for checkpointing simulations

checkpointFreq
Saving frequency for checkpointing

checkTimeStep
If TRUE, warnings will be issued if TimeStep is too short.

verboseNested
Should looped runs of simulateWithPriors be verbose?

Value

Function simulateWithPriors returns a vector of trueFreeValues, the true generating parameters used in the simulation (a set of values as long as the number of freely varying parameters), concatenated with a set of summary statistics for the simulation. Function parallelSimulateWithPriors returns a matrix of such vectors bound together, with each row representing a different simulation. By default, both functions also assign a logical vector named freevector, indicating the total number of parameters and which parameters are freely-varying (have TRUE values), as an attribute of the output.

Note

The simulateWithPriors functions are effectively the engine that powers the doRun functions, while the doSimulation function is the pistons within the simulateWithPriors engine. In general, most users will just drive the car - they will just use doRun, but some users may want to use simulateWithPriors or doSimulation to do various simulations.

Examples

set.seed(1)
tree <- rcoal(20)
# get realistic edge lengths
tree$edge.length <- tree$edge.length*20

# example simulation

# NOTE: the example analyses involve too few simulations, with coarse time-units
    # - all for the sake of examples that reasonably test the functions

simData <- simulateWithPriors(phy = tree,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 100000,
     freevector = NULL,
     giveUpAttempts = 10,
     verbose = TRUE)

#> Warning: You have only 15 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 0.0011

simData

#>                                                                       
#>  9.914560e+00  9.773958e-02 -4.507873e+01  1.495949e+00  9.415746e+01 
#>                                                                       
#> -4.507873e+01  1.495949e+00  1.000000e+00  9.615746e+01 -4.452508e+01 
#>                                                                       
#>  2.863875e+00  4.158169e-01  9.505017e+01 -4.507873e+01  1.495950e+00 
#>                                                                       
#>  2.060040e-08  9.615746e+01 -6.975056e+01  1.435011e+02  1.089262e+01 
#>                                                                       
#>  1.904446e+01 -5.267234e+00  6.592202e+01  1.358663e+01 -4.270773e+00 
#>                                                                       
#> -2.514268e+01  3.786493e+00  6.642739e+00  5.236797e+00 -8.038566e+00 
#>                                                                       
#>  6.336197e+00  5.691886e+00  9.118213e-01  2.400947e+00 -1.488146e-01 
#>                                                                       
#> -1.666962e+00 -1.365898e+00  5.070571e-01  2.159636e-01  3.281078e-01 
#>                                                                    21 
#>  5.745538e-02  4.253937e-01  1.305710e-01 -1.775722e-02  7.084204e+00 
#>            22            23            24            25            26 
#>  9.834346e+00  1.043319e+01  1.271108e+01  9.067655e+00  1.241538e+01 
#>            27            28            29            30            31 
#>  1.544848e+01  1.662789e+01  5.375870e+00 -3.923583e+00 -3.934039e+00 
#>            32            33            34            35            36 
#>  1.535083e+01  1.801773e+01  1.212858e+01  1.624807e+01  1.764738e+01 
#>            37            38            39                             
#>  1.866050e+01  1.821376e+01  3.360913e+00 -2.667103e-01  4.091029e+00 
#>                                                                       
#>  5.007883e+00  8.238364e+00  4.578631e+00  8.393363e+00  1.278713e+01 
#>                                                                       
#>  1.433891e+01  2.885739e+00 -5.745323e+00 -5.726866e+00  1.349522e+01 
#>                                                                       
#>  1.638063e+01  1.052149e+01  1.528591e+01  1.678003e+01  1.784245e+01 
#>                                                                       
#>  1.742704e+01  2.872938e+00  1.443512e+01  1.557766e+01  1.585850e+01 
#>                                                                       
#>  1.718381e+01  1.355668e+01  1.643740e+01  1.810983e+01  1.891688e+01 
#>                                                                       
#>  7.866001e+00 -2.101843e+00 -2.141213e+00  1.720645e+01  1.965483e+01 
#>                                                                       
#>  1.373567e+01  1.721024e+01  1.851474e+01  1.947855e+01  1.900048e+01 
#>               
#>  3.848889e+00 
#> attr(,"freevector")
#>  starting_1 intrinsic_1 extrinsic_1 
#>        TRUE        TRUE       FALSE 

simDataParallel <- parallelSimulateWithPriors(
  nrepSim = 2, multicore = FALSE, coreLimit = 1,
  phy = tree,
  intrinsicFn = brownianIntrinsic,
  extrinsicFn = nullExtrinsic,
  startingPriorsFns = "normal",
  startingPriorsValues = matrix(c(mean(simChar[, 1]), sd(simChar[, 1]))),
  intrinsicPriorsFns = c("exponential"),
  intrinsicPriorsValues = matrix(c(10, 10), nrow = 2, byrow = FALSE),
  extrinsicPriorsFns = c("fixed"),
  extrinsicPriorsValues = matrix(c(0, 0), nrow = 2, byrow = FALSE),
  generation.time = 100000,
  checkpointFile = NULL, checkpointFreq = 24,
  verbose = TRUE, freevector = NULL, taxonDF = NULL)

#> You have only 15 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 0.0011

#> Using 1 core(s) for simulations 
#> 

#> Doing simulations: 

simDataParallel

#>                                                                        
#> result.1 9.909174 0.005462155  13.50491 0.004272301 -23.00981  13.50491
#> result.2 9.191764 0.073686613 -41.36862 1.032259719  86.73724 -41.36862
#>                                                                           
#> result.1 0.004272301 1 -21.00981  13.50967 0.004032043 1.0785437 -21.01934
#> result.2 1.032259719 1  88.73724 -41.20150 1.491355768 0.6135012  88.40300
#>                                                                           
#> result.1  13.50491 0.004272307 1.602096e-07 -21.00981  -7.858761  19.71752
#> result.2 -41.36862 1.032260125 3.659933e-08  88.73725 -63.287317 130.57463
#>                                                                              
#> result.1 9.617276 10.10798 8.9845455  0.1352395 9.752255  9.724531  0.5595722
#> result.2 8.150636 19.63719 0.5547008 34.5410321 6.554852 18.271090 17.3750301
#>                                                                              
#> result.1  0.9586729  0.2860417 -0.03955481  -0.1808732  0.3495666 -0.07252331
#> result.2 -8.0401147 -1.4254631  6.32853103 -10.7531647 -1.4693653 -1.42807789
#>                                                                            
#> result.1  0.1559201 -0.009738236 -0.1407142 0.1064007 -0.1335336 0.06931171
#> result.2 -1.0363169  0.637397083 -0.4775456 1.1048244  1.2715129 1.96083132
#>                                                                      
#> result.1 -0.01828816  0.03484687  0.01573799 -0.04255751 -0.001271947
#> result.2  0.59347392 -0.33602950 -0.43103995 -0.57876175 -0.468884499
#>                            21       22       23       24       25       26
#> result.1 0.01236692  9.588144 9.526937 9.405694 9.725431 9.183392 9.767720
#> result.2 0.18865402 10.892917 8.992412 9.139743 5.639294 9.739513 4.039729
#>                27       28       29        30        31       32       33
#> result.1 9.775639 9.766855 9.017313  9.833132  9.837173 9.986836 9.803281
#> result.2 6.749080 6.584984 5.852413 18.499929 18.522501 1.731726 6.130150
#>                 34         35       36       37       38       39         
#> result.1  9.160744 10.0691838 9.685017 9.844816 9.783217 8.996847 9.195305
#> result.2 14.993270  0.9300788 6.858116 6.857042 5.109243 4.857660 4.786627
#>                                                                        
#> result.1 9.220010 9.115762 9.486405 8.943495 9.5527807 9.633415 9.64453
#> result.2 4.221528 4.633022 1.923875 6.010551 0.6987037 4.538336 4.68356
#>                                                                             
#> result.1 8.884238  9.735777  9.741363 9.8876708 9.715793  9.07486 10.0177650
#> result.2 3.783900 16.986638 17.033228 0.1902968 4.770235 13.65829  0.1308248
#>                                                                           
#> result.1 9.638665 9.801099 9.741174 8.970769  9.980982  9.833864  9.695627
#> result.2 6.137620 6.177500 4.455727 4.452306 16.999207 13.763295 13.646464
#>                                                                          
#> result.1 9.964457  9.423288 9.982660 9.917864 9.889180 9.150387  9.930487
#> result.2 9.354714 13.468475 7.380755 8.959825 8.486409 7.920926 20.013220
#>                                                                            
#> result.1  9.932983 10.086001 9.890769  9.246628 10.120602 9.731369 9.888533
#> result.2 20.011774  3.273155 7.490065 16.328253  1.729333 7.578613 7.536584
#>                           
#> result.1 9.825259 9.022925
#> result.2 5.762758 5.263013
#> attr(,"freevector")
#>  starting_1 intrinsic_1 extrinsic_1 
#>        TRUE        TRUE       FALSE