Simulate data for initial TreEvo analysis

The simulateWithPriors function pulls parameters from prior distributions and conducts a single simulation of continuous trait evolution (using doSimulation functions), returning useful summary statistics for ABC. parallelSimulateWithPriors is a wrapper function for simulateWithPriors that allows for multithreading and checkpointing. This family of functions is mostly used as internal components, generating simulations within ABC analyses using the doRun functions. See Note below.

simulateWithPriors(phy = NULL, intrinsicFn, extrinsicFn, startingPriorsFns,
  startingPriorsValues, intrinsicPriorsFns, intrinsicPriorsValues,
  extrinsicPriorsFns, extrinsicPriorsValues, generation.time = 1000,
  TreeYears = max(branching.times(phy)) * 1e+06, timeStep = NULL,
  giveUpAttempts = 10, verbose = FALSE, checks = TRUE, taxon.df = NULL,
  freevector = NULL, niter.brown = 25, niter.lambda = 25,
  niter.delta = 25, niter.OU = 25, niter.white = 25)

parallelSimulateWithPriors(nrepSim, multicore, coreLimit, phy, intrinsicFn,
  extrinsicFn, startingPriorsFns, startingPriorsValues, intrinsicPriorsFns,
  intrinsicPriorsValues, extrinsicPriorsFns, extrinsicPriorsValues,
  generation.time = 1000, TreeYears = max(branching.times(phy)) * 1e+06,
  timeStep = NULL, checkpointFile = NULL, checkpointFreq = 24,
  verbose = FALSE, freevector = NULL, taxon.df = NULL,
  giveUpAttempts = 10, niter.brown = 25, niter.lambda = 25,
  niter.delta = 25, niter.OU = 25, niter.white = 25)

Arguments

phy
A phylogenetic tree, in package ape's phylo format.

intrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, regardless of the trait values of other taxa.

extrinsicFn
Name of (previously-defined) function that governs how traits evolve within a lineage, based on their own ('internal') trait vlaue and the trait values of other taxa.

startingPriorsFns
Vector containing names of prior distributions to use for root states: can be one of "fixed", "uniform", "normal",

startingPriorsValues
Matrix with number of columns equal to the number of states (characters) at root and number of rows equal to two (representing two parameters to pass to prior distribution).

intrinsicPriorsFns
Vector containing names of prior distributions to use for intrinsic function parameters: can be one of "fixed", "uniform", "normal",

intrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the intrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

extrinsicPriorsFns
Vector containing names of prior distributions to use for extrinsic function parameters: can be one of "fixed", "uniform", "normal",

extrinsicPriorsValues
Matrix with number of columns equal to the number of parameters to pass to the extrinsic function and number of rows equal to two (representing two parameters to pass to prior distribution).

timeStep
This value corresponds to the number of discrete evolutionary events ('generations') simulated per time-unit along branches. Typically NULL by default and determined internally as follows: timeStep = generation.time / TreeYears. Can be provided a value as an alternative to using arguments generation.time and TreeYears, which would then be overridden.

giveUpAttempts
Value for when to stop the analysis if NA values are present.

verbose
If TRUE, gives messages about how the simulation is progessing via print.

checks
If TRUE, checks inputs for consistency. This activity is skipped (checks = FALSE) when run in parallel by parallelSimulateWithPriors, and instead is only checked once. This argument also controls whether simulateWithPriors assigns freevector as an attribute to the output produced.

taxon.df
A data.frame containing data on nodes (both tips and internal nodes) output by various internal functions. Can be supplied as input to spead up repeated calculations, but by default is NULL, which instead forces a calculation from input phy.

freevector
A logical vector (with length equal to the number of parameters), indicating free (TRUE) and fixed (FALSE) parameters.

niter.brown
Number of random starts for the Brownian Motion (BM) model (minimum of 2).

niter.lambda
Number of random starts for the lambda model (minimum of 2).

niter.delta
Number of random starts for the delta model (minimum of 2).

niter.OU
Number of random starts for the Ornstein-Uhlenbeck (OU) model (minimum of 2).

niter.white
Number of random starts for the white noise model (minimum of 2).

nrepSim
Number of replicated simulations to run.

multicore
Whether to use multicore, default is FALSE. If TRUE, one of two suggested packages must be installed, either doMC (for UNIX systems) or doParallel (for Windows), which are used to activate multithreading. If neither package is installed, this function will fail if multicore=TRUE.

coreLimit
Maximum number of cores to be used.

checkpointFile
Optional file name for checkpointing simulations

checkpointFreq
Saving frequency for checkpointing

Value

Function simulateWithPriors returns a vector of trueFreeValues, the true generating parameters used in the simulation (a set of values as long as the number of freely varying parameters), concatenated with a set of summary statistics for the simulation. Function parallelSimulateWithPriors returns a matrix of such vectors bound together, with each row representing a different simulation. By default, both functions also assign a logical vector named freevector, indicating the total number of parameters and which parameters are freely-varying (have TRUE values), as an attribute of the output.

Note

The simulateWithPriors functions are effectively the engine that powers the doRun functions, while the doSimulation functions are the pistons within the simulateWithPriors engine. In general, most users will just drive the car - they will just use doRun, but some users may want to use simulateWithPriors or doSimulation functions to do various simulations.

Examples

simPhy<-rcoal(30)

# example simulation

# NOTE: the example analyses involve too few simulations, with coarse time-units
        # - all for the sake of examples that reasonably test the functions

simData<-simulateWithPriors(phy=simPhy,
  intrinsicFn=brownianIntrinsic,
  extrinsicFn=nullExtrinsic,
  startingPriorsFns="normal",
  startingPriorsValues=matrix(c(mean(simChar[,1]), sd(simChar[,1]))),
  intrinsicPriorsFns=c("exponential"),
  intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE),
  extrinsicPriorsFns=c("fixed"),
  extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
         freevector=NULL,
         giveUpAttempts=10,
         verbose=FALSE,
         niter.brown=25, niter.lambda=25, niter.delta=25, niter.OU=25, niter.white=25)

#> Warning: You have only 0 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 2.8e-05 or at the very least 0.00046

#> .

simData

#>   [1]  9.908322e+00  2.370446e-01 -6.311727e+01  3.203581e+01  1.302345e+02
#>   [6] -6.311727e+01  3.203582e+01  1.000000e+00  1.322345e+02 -6.307298e+01
#>  [11]  3.649688e+01  8.363340e-01  1.321460e+02 -6.311727e+01  3.203581e+01
#>  [16]  3.238837e-15  1.322345e+02 -8.918667e+01  1.823733e+02  9.938120e+00
#>  [21]  1.780754e+01 -7.095555e-01  2.314555e+01  1.036164e+01  9.860627e+00
#>  [26] -1.063149e-02  1.557755e+01  1.633588e+01  5.940589e+00  3.022124e+00
#>  [31] -4.969269e+00  2.929633e-01  6.313068e-01  3.322295e+00 -5.851893e+00
#>  [36] -5.497438e-01 -2.169574e+00  5.120780e-01 -2.946518e-01 -1.188859e+00
#>  [41]  2.586752e-01  3.439354e+00  2.356067e+00  1.064898e+00 -6.607441e-01
#>  [46] -1.096689e+00 -1.090193e+00  7.406959e-01 -9.737233e-01 -4.197337e-01
#>  [51]  5.820778e-01 -1.123243e-02 -4.057629e-01 -9.499417e-02  9.869027e+00
#>  [56]  9.869729e+00  8.945021e+00  1.225310e+01  1.227532e+01  1.006827e+01
#>  [61]  3.682402e+00  9.712909e+00  1.326162e+01  1.104493e+01  3.089535e+00
#>  [66]  8.956391e+00  9.164049e+00  1.334754e+01  1.243466e+01  1.717964e+01
#>  [71]  1.991637e+00  1.191075e+01  1.217663e+01  1.845052e+00  1.002733e+01
#>  [76]  1.215520e+01  9.842525e+00  1.087429e+01  1.042108e+01  1.393433e+01
#>  [81]  1.422211e+01  2.267925e+00  1.144971e+01  3.380295e+01  2.923684e+01
#>  [86]  2.843090e+01  2.694402e+01  2.681751e+01  2.314022e+01  2.435894e+01
#>  [91]  2.265187e+01  2.463125e+01  2.234430e+01  2.293940e+01  1.745364e+01
#>  [96]  1.650716e+01  1.712911e+01  1.558309e+01  1.560995e+01  1.507591e+01
#> [101]  9.928157e+00  8.809562e+00  1.165089e+01  9.689760e+00  8.628801e+00
#> [106]  8.944470e+00  8.139791e+00  7.317491e+00  6.098688e+00  5.546326e+00
#> [111]  3.177931e+00  2.300708e+00
#> attr(,"freevector")
#> [1]  TRUE  TRUE FALSE

simDataParallel<-parallelSimulateWithPriors(
  nrepSim=2, multicore=FALSE, coreLimit=1,
  phy=simPhy,
  intrinsicFn=brownianIntrinsic,
  extrinsicFn=nullExtrinsic,
  startingPriorsFns="normal",
  startingPriorsValues=matrix(c(mean(simChar[,1]), sd(simChar[,1]))),
  intrinsicPriorsFns=c("exponential"),
  intrinsicPriorsValues=matrix(c(10, 10), nrow=2, byrow=FALSE),
  extrinsicPriorsFns=c("fixed"),
  extrinsicPriorsValues=matrix(c(0, 0), nrow=2, byrow=FALSE),
  checkpointFile=NULL, checkpointFreq=24,
  verbose=FALSE,
  freevector=NULL, taxon.df=NULL,
  niter.brown=25, niter.lambda=25, niter.delta=25, niter.OU=25, niter.white=25)

#> You have only 0 timeSteps on the shortest branch in this dataset but should probably have a lot more if you expect change on this branch. Please consider decreasing timeStep to no more than 2.8e-05 or at the very least 0.00046

#> Using 1 core(s) for simulations 
#> 

#> Doing simulations: 

#> .

#> .

simDataParallel

#>               [,1]       [,2]       [,3]     [,4]      [,5]       [,6]
#> result.1  9.012181 0.01749904   4.768877 0.346841 -5.537755   4.848214
#> result.2 10.164907 0.03753726 -11.189375 1.005007 26.378750 -10.726993
#>               [,7]      [,8]      [,9]      [,10]     [,11]     [,12]     [,13]
#> result.1 0.3283678 0.9983297 -3.696427   4.891727 0.4344880 0.7339633 -3.783455
#> result.2 0.8889677 0.9964379 27.453985 -10.310039 0.5376023 2.3490758 26.620079
#>               [,14]    [,15]        [,16]     [,17]     [,18]    [,19]
#> result.1   4.768877 0.346841 3.472303e-14 -3.537755 -26.20687 56.41373
#> result.2 -10.260960 1.334283 1.707510e+00 26.521920 -22.12092 48.24183
#>              [,20]    [,21]    [,22]     [,23]     [,24]    [,25]      [,26]
#> result.1  8.957238 10.01219 7.863766 0.3475485  9.025509 8.633139 -0.4603334
#> result.2 10.749448 11.71010 9.546204 0.2646763 10.733902 9.911156 -0.6998058
#>              [,27]      [,28]      [,29]      [,30]      [,31]      [,32]
#> result.1  1.729439 -1.1217029 -0.5581929 -0.4217087  0.6768277 -0.1866917
#> result.2 -1.722637 -0.4796156  0.6256921  0.3277753 -0.2292595  0.1210150
#>               [,33]      [,34]      [,35]      [,36]      [,37]      [,38]
#> result.1 -0.5018875 -0.2311755 -0.6669822  0.1585616  0.1636599 0.20919968
#> result.2  0.3408300 -0.9460945 -1.1446692 -0.9661201 -0.3700911 0.02124094
#>               [,39]      [,40]      [,41]       [,42]      [,43]       [,44]
#> result.1 -0.2040562 -0.1699641  0.2204416  0.07821815 0.04943021 -0.03098165
#> result.2 -0.1790219  0.2286483 -0.3060061 -0.17798519 0.11091494 -0.27609049
#>                [,45]       [,46]      [,47]       [,48]       [,49]      [,50]
#> result.1 -0.06403077  0.06637333 -0.1147947 -0.06090384 -0.04942207 0.11301796
#> result.2  0.36870298 -0.24087105 -0.1338208  0.32781613 -0.13046890 0.05109094
#>                [,51]      [,52]       [,53]       [,54]     [,55]     [,56]
#> result.1 0.066892552 0.04253411 -0.05345159  0.01463165  9.000579  9.033537
#> result.2 0.005761313 0.09978652 -0.04080346 -0.06777156 10.470034 10.519999
#>              [,57]     [,58]     [,59]     [,60]     [,61]     [,62]     [,63]
#> result.1  8.950646  9.338745  9.380202  8.595108  9.027861  8.618744  9.480998
#> result.2 10.652663 10.318171 10.283229 10.820690 11.012294 10.818479 10.190439
#>              [,64]     [,65]     [,66]     [,67]     [,68]     [,69]     [,70]
#> result.1  8.427187  9.014593  8.829602  8.829707  8.447749  9.617531  9.252872
#> result.2 10.885758 11.069705 10.696130 10.732243 11.227133 10.513865 10.514351
#>              [,71]     [,72]     [,73]     [,74]     [,75]     [,76]     [,77]
#> result.1  9.134602  9.587975  9.590974  9.060864  8.008546  9.595476  8.009054
#> result.2 11.064425 10.643881 10.677879 10.749120 10.550272 10.720605 10.507191
#>             [,78]     [,79]     [,80]     [,81]    [,82]     [,83]    [,84]
#> result.1  9.56184  7.962223  8.358321  8.552235  8.90786  9.540072 0.924617
#> result.2 10.63187 10.631497 11.250492 11.421719 11.67747 10.760584 1.573916
#>              [,85]     [,86]     [,87]     [,88]     [,89]     [,90]     [,91]
#> result.1 0.7997188 0.7776755 0.7370061 0.7335456 0.6329606 0.6662964 0.6196028
#> result.2 1.3613094 1.3237865 1.2545577 1.2486670 1.0774478 1.1341931 1.0547096
#>              [,92]     [,93]     [,94]     [,95]     [,96]     [,97]     [,98]
#> result.1 0.6737465 0.6111902 0.6274695 0.4774146 0.4515257 0.4685379 0.4262491
#> result.2 1.1468749 1.0403894 1.0681006 0.8126719 0.7686028 0.7975617 0.7255762
#>              [,99]    [,100]    [,101]    [,102]    [,103]    [,104]    [,105]
#> result.1 0.4269842 0.4123770 0.2715681 0.2409708 0.3186918 0.2650471 0.2360264
#> result.2 0.7268274 0.7019625 0.4622727 0.4101890 0.5424883 0.4511725 0.4017724
#>             [,106]    [,107]    [,108]    [,109]    [,110]    [,111]     [,112]
#> result.1 0.2446610 0.2226503 0.2001578 0.1668194 0.1517105 0.0869270 0.06293203
#> result.2 0.4164705 0.3790032 0.3407155 0.2839659 0.2582469 0.1479702 0.10712511
#> attr(,"freevector")
#> [1]  TRUE  TRUE FALSE