Summarize Posterior Distribution for a Free Parameter

This function summarizes the distribution of parameter estimates from the posterior of an ABC analysis sing the doRun functions in TreEvo, for all freely varying parameters. Only the final generation is considered. This summary includes the mean, standard deviation and Highest Posterior Density (at a 0.8 alpha) for each parameter.

summarizePosterior(particleDataFrame, alpha = 0.8, coda = FALSE,
  verboseMultimodal = TRUE, ...)

Arguments

particleDataFrame
A particleDataFrame object, as found among the output from doRun functions.

alpha
The threshold used for defining the highest density frequency cut-off. If the highest density interval is applied to a Bayesian MCMC posterior sample, then the interval is effectively calculated for this value as a posterior probability density.

coda
Default is FALSE. If TRUE, unimodal highest density regions will instead be calculated using HPDinterval from package coda, which is similar to the function quantile in that it calculates only a single interval.

verboseMultimodal
If TRUE, the function will print a message indicating when the inferred highest density interval is discontinuous, and thus likely reflects that the supplied data is multimodal.

...
Additional arguments passed to density. A user may want to mess with this to adjust bandwidth, et cetera.

Value

Returns a list, wherein each element of the list is secondary list containing the weighted mean, standard deviation, and a matrix giving the highest density intervals (e.g. the highest posterior density intervals). Because posterior estimates of parameter values may be multimodal, multiple sets of bounds may be reported for complex posterior distributions, which each constitute one row of the output matrix. See highestDensityInterval for details.

See also

Examples

# example with output from doRun_prc
data(simRunExample)

# alpha = 0.95
summarizePosterior(results[[1]]$particleDataFrame, alpha = 0.95)

#> $starting_1
#> $starting_1$mean
#> [1] 9.661602
#> 
#> $starting_1$sd
#> [1] 0.1985405
#> 
#> $starting_1$HPD
#>      LowerBound_alpha=0.95 UpperBound__alpha=0.95
#> [1,]              9.292509                10.0967
#> 
#> 
#> $intrinsic_1
#> $intrinsic_1$mean
#> [1] 0.04162821
#> 
#> $intrinsic_1$sd
#> [1] 0.02389411
#> 
#> $intrinsic_1$HPD
#>      LowerBound_alpha=0.95 UpperBound__alpha=0.95
#> [1,]          -0.006581125             0.08965948
#> 
#> 

# you might be tempted to use alphas like 95%, but with bayesian statistics
# we often don't sample the distribution well enough to know
# its shape to exceeding detail. alpha = 0.8 may be more reasonable.
summarizePosterior(results[[1]]$particleDataFrame, alpha = 0.8)

#> Inferred highest density intervals are discontinuous, suggesting data may have multiple modes

#> $starting_1
#> $starting_1$mean
#> [1] 9.661602
#> 
#> $starting_1$sd
#> [1] 0.1985405
#> 
#> $starting_1$HPD
#>      LowerBound_alpha=0.8 UpperBound__alpha=0.8
#> [1,]             9.368550              9.845534
#> [2,]             9.969964             10.050613
#> 
#> 
#> $intrinsic_1
#> $intrinsic_1$mean
#> [1] 0.04162821
#> 
#> $intrinsic_1$sd
#> [1] 0.02389411
#> 
#> $intrinsic_1$HPD
#>      LowerBound_alpha=0.8 UpperBound__alpha=0.8
#> [1,]          0.005704909            0.07386315
#> 
#> 

# or even better, for coverage purposes, maybe 0.5
summarizePosterior(results[[1]]$particleDataFrame, alpha = 0.5)

#> $starting_1
#> $starting_1$mean
#> [1] 9.661602
#> 
#> $starting_1$sd
#> [1] 0.1985405
#> 
#> $starting_1$HPD
#>      LowerBound_alpha=0.5 UpperBound__alpha=0.5
#> [1,]             9.506806              9.725712
#> 
#> 
#> $intrinsic_1
#> $intrinsic_1$mean
#> [1] 0.04162821
#> 
#> $intrinsic_1$sd
#> [1] 0.02389411
#> 
#> $intrinsic_1$HPD
#>      LowerBound_alpha=0.5 UpperBound__alpha=0.5
#> [1,]           0.02501153            0.06245469
#> 
#>